% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiweek.R
\name{assert_date_in_epiweek}
\alias{assert_date_in_epiweek}
\title{assert that a date corresponds to a given
epidemiological week and year, according to the
given epiweek standard.}
\usage{
assert_date_in_epiweek(
  date,
  expected_epiweek,
  expected_epiyear,
  epiweek_standard
)
}
\arguments{
\item{date}{Vector of dates to validate.}

\item{expected_epiweek}{Vector of expected epiweek values.}

\item{expected_epiyear}{Vector of expected epiyear values.}

\item{epiweek_standard}{Epiweek standard to use.
One of \code{"USA"} or \code{"MMWR"} (USA / MMWR
epiweek, starts on Sunday) or \code{"ISO"} (ISO
week, starts on Monday). Not case-sensitive.}
}
\value{
Nothing on success, or raise an error if the assertion
fails.
}
\description{
assert that a date corresponds to a given
epidemiological week and year, according to the
given epiweek standard.
}
\examples{
assert_date_in_epiweek("2025-01-02", 1, 2025, "USA")
assert_date_in_epiweek(
  c("2024-12-27", "2025-01-05"),
  c(52, 1),
  c(2024, 2025),
  "ISO"
)

tryCatch(
  assert_date_in_epiweek(
    "2024-12-01", 52, 2024, "ISO"
  ),
  error = \(e) print(e)
)

}
