% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkers.R
\name{assert_hubverse_output_types}
\alias{assert_hubverse_output_types}
\title{Assert that a vector consists only
of valid hubverse output types}
\usage{
assert_hubverse_output_types(x)
}
\arguments{
\item{x}{vector to check}
}
\value{
\code{x}, invisibly, on assertion success, or throw an error.
}
\description{
Wraps \code{\link[checkmate:assert_names]{checkmate::assert_names()}}.
}
\examples{

assert_hubverse_output_types(
  c("sample", "sample", "quantile", "pmf"))

tryCatch(
  assert_hubverse_output_types(
    c("sample", "sample", "sample_a")
  ),
  error = \(e) print(e)
)

}
