% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_bottom_up.R
\name{bottom_up_aggregation}
\alias{bottom_up_aggregation}
\title{Bottom up aggregation of base forecasts}
\usage{
bottom_up_aggregation(
  base_forecasts,
  cp,
  ndraws,
  draw_col = ".draw",
  date_col = "date",
  value_to_aggregate_col = "hosp",
  rank_quantity_col = "hosp",
  location_col = "location",
  aggregated_location_name = "aggregate"
)
}
\arguments{
\item{base_forecasts}{The base forecasts tibble}

\item{cp}{Base forecast rank sampling copula.}

\item{ndraws}{Number of draws from rank sampling copula}

\item{draw_col}{String name of column for draws.}

\item{date_col}{String name of column for dates.}

\item{value_to_aggregate_col}{String name of column to make bottom-up
aggregation on.}

\item{rank_quantity_col}{String name of column to rank draw/location
pair on.}

\item{location_col}{String name of the location column.}

\item{aggregated_location_name}{Name for the aggregated location.
Default \code{"aggregate"}.}
}
\value{
A tibble of 1,..., \code{ndraws} aggregated forecasts.
}
\description{
Bottom up aggregation of base forecasts
}
