% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorize_vector.R
\name{categorize_vector}
\alias{categorize_vector}
\title{Catgeorize entries in a numeric vector according
to a list of breakpoints and labels.}
\usage{
categorize_vector(
  values,
  break_sets,
  label_sets,
  include.lowest = TRUE,
  order = TRUE,
  right = TRUE,
  ...
)
}
\arguments{
\item{values}{Vector of values to categorize.}

\item{break_sets}{list of vectors of category
breakpoints, one for each entry of \code{values},
or a single set of category breakpoints to use
for all \code{values}.}

\item{label_sets}{list of vectors of labels to associate
with the the corresponding breakpoints in \code{break_sets},
one for each entry of \code{break_sets} or a single set of
labels to use for all \code{break_sets}.}

\item{include.lowest}{Passed to \code{\link[base:cut]{base::cut()}}. Default \code{TRUE}.}

\item{order}{Passed to \code{\link[base:cut]{base::cut()}}. Default \code{TRUE}.}

\item{right}{Passed to \code{\link[base:cut]{base::cut()}}. Default \code{TRUE}.}

\item{...}{Additional keyword arguments passed to
\code{\link[base:cut]{base::cut()}}.}
}
\value{
A categorized vector, as vector of ordered
factors.
}
\description{
Vectorized version of \code{\link[base:cut]{base::cut()}} that
allows each entry in the vector to use different
break points and labels.
}
