% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round_epiweek.R
\name{ceiling_isoweek}
\alias{ceiling_isoweek}
\title{Get the final date in an ISO week}
\usage{
ceiling_isoweek(date)
}
\arguments{
\item{date}{A date.}
}
\value{
The date that ends the epidemiological week, as a
\code{\link[lubridate:date]{lubridate::date()}}.
}
\description{
Given any date from an ISO week ("isoweek") return
the date of the final day of that isoweek (a Sunday).
}
\examples{
ceiling_isoweek(as.Date("2025-03-13")) # rounded to Sunday
ceiling_isoweek(as.Date("2025-03-16")) # this is already Sunday
}
