% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round_epiweek.R
\name{ceiling_mmwr_epiweek}
\alias{ceiling_mmwr_epiweek}
\title{Get the final date in a US/MMWR epidemiological week}
\usage{
ceiling_mmwr_epiweek(date)
}
\arguments{
\item{date}{A date.}
}
\value{
The date that ends the epidemiological week, as a
\code{\link[lubridate:date]{lubridate::date()}}.
}
\description{
Given any date from a US/MMWR epidemiological week ("epiweek") return
the date of the final day of the epiweek (a Saturday).
}
\examples{
ceiling_mmwr_epiweek(as.Date("2025-03-13")) # rounded to Saturday
ceiling_mmwr_epiweek(as.Date("2025-03-15")) # this is already Saturday
}
