% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{centered_eight}
\alias{centered_eight}
\title{Eight Schools InferenceData DataFrame}
\format{
\subsection{\code{centered_eight}}{

A data frame with 8 rows and 62 columns:
}
}
\source{
\url{https://github.com/arviz-devs/arviz/blob/main/arviz/data/example_data/code/centered_eight/centered_eight.ipynb}
}
\usage{
centered_eight
}
\description{
A subset of data from ArviZ's centered_eight dataset
}
\keyword{datasets}
