% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_bottom_up.R
\name{copula2tbl}
\alias{copula2tbl}
\title{Convert Copula Samples to a Tibble}
\usage{
copula2tbl(i, u_mat, location_names, location_col)
}
\arguments{
\item{i}{The index of the sample.}

\item{u_mat}{The matrix of Copula samples.}

\item{location_names}{The names of the locations.}

\item{location_col}{String name of the location column.}
}
\value{
A tibble containing the Copula samples and location names.
}
\description{
This function converts a matrix of copula samples to a tibble format.
}
