% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_bottom_up.R
\name{count_trajectories}
\alias{count_trajectories}
\title{Count the number of trajectories per location}
\usage{
count_trajectories(base_forecasts, location_col, date_col)
}
\arguments{
\item{base_forecasts}{The base forecast dataset.}

\item{location_col}{The name of the column containing the location
information.}

\item{date_col}{The name of the column containing the date information.}
}
\value{
A tibble with the number of samples per forecast, grouped
by location.
}
\description{
This function takes a base forecast dataset and counts the number of
trajectories per location. In this context a trajectory is a sequence
of forecast values (i.e. rows in base forecast) with the same
draw id and location along with ascending date values.
}
