% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_table_for_scoring.R
\name{create_table_for_scoring}
\alias{create_table_for_scoring}
\title{Create a table for scoring hub models
with \code{scoringutils}}
\usage{
create_table_for_scoring(hub_path)
}
\arguments{
\item{hub_path}{Local path to forecast hub.}
}
\value{
Table with submitted model forecasts
and truth data, as a \code{\link[tibble:tibble]{tibble::tibble()}}
}
\description{
Create a table for scoring models
submitted to a forecast hub using
\code{scoringutils} by doing needed formatting
and attaching truth data.
}
\details{
Requires a local version of the
forecast hub at \code{hub_path}.
}
