% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_data_cdc_gov.R
\name{data_cdc_gov_base_query}
\alias{data_cdc_gov_base_query}
\title{Construct a base SOQL query pointing at at data.cdc.gov endpoint.}
\usage{
data_cdc_gov_base_query(dataset_id)
}
\arguments{
\item{dataset_id}{Dataset ID, as a string.}
}
\value{
A \code{\link[=soql:soql]{soql:soql()}} object pointing at the dataset's JSON
API endpoint on \code{data.cdc.gov}.
}
\description{
Construct a base SOQL query pointing at at data.cdc.gov endpoint.
}
\examples{

data_cdc_gov_dataset_id("nssp_prop_ed_visits") |>
  data_cdc_gov_endpoint() |>
  data_cdc_gov_base_query()

}
