% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_data_cdc_gov.R
\name{data_cdc_gov_dataset_id}
\alias{data_cdc_gov_dataset_id}
\title{Retrieve the dataset ID for a supported \code{data.cdc.gov} dataset}
\usage{
data_cdc_gov_dataset_id(dataset_name)
}
\arguments{
\item{dataset_name}{Internal forecasttools name for the dataset.
See \link{data_cdc_gov_datasets}.}
}
\value{
The dataset id, as a string.
}
\description{
Retrieve the dataset ID for a supported \code{data.cdc.gov} dataset
}
\examples{
data_cdc_gov_dataset_id("nhsn_hrd_prelim")

data_cdc_gov_dataset_id("nssp_prop_ed_visits")

}
