% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_data_cdc_gov.R
\name{data_cdc_gov_dataset_lookup}
\alias{data_cdc_gov_dataset_lookup}
\title{Look up rows of the data.cdc.gov dataset table
corresponding to the entries of a given
vector.}
\usage{
data_cdc_gov_dataset_lookup(dataset, input_format)
}
\arguments{
\item{dataset}{vector of dataset keys or ids}

\item{input_format}{format in which the input \code{dataset}
vector is coded. One of \code{"key"} or \code{"id"}.}
}
\value{
A \code{\link[tibble:tibble]{tibble}} with the
corresponding rows of the \link{data_cdc_gov_dataset_table}
matching the location vector (with repeats possible).
Returns \code{NA} rows where matches cannot be found.
}
\description{
Look up rows of the data.cdc.gov dataset table
corresponding to the entries of a given
vector.
}
\examples{

data_cdc_gov_dataset_lookup(
  c("nhsn_hrd_prelim", "nhsn_hrd_final"),
  "key")

data_cdc_gov_dataset_lookup(
  c("rdmq-nq56", "ua7e-t2fy", "ua7e-t2fy"),
  "id")

data_cdc_gov_dataset_lookup(
  c("rdmq-nq56", "unavailable", "ua7e-t2fy"),
  "id",
  strict = FALSE)

}
