% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_data_cdc_gov.R
\name{data_cdc_gov_dataset_table}
\alias{data_cdc_gov_dataset_table}
\title{Dataset IDs and metadata for \code{data.cdc.gov} datasets}
\usage{
data_cdc_gov_dataset_table
}
\description{
\code{nhsn_hrd_final}, \code{nhsn_hrd_prelim}: NHSN Hospital Respiratory Data
(HRD).
Epiweekly-resolution incident hospital admissions
timeseries for COVID-19, Influenza, and RSV.
\code{nhsn_hrd_final} points to the most recent final (Friday) weekly data
release. \code{nhsn_hrd_prelim} points the most recent preliminary (Wednesday)
weekly release.
}
\details{
\code{nssp_prop_ed_visits}: National Syndromic Surveillance
Program (NSSP) data on proportion of incident emergency
deparment (ED) visits due to COVID-19, Influenza, and RSV.
}
\seealso{
\code{\link[=data_cdc_gov_dataset_id]{data_cdc_gov_dataset_id()}}
}
