% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_data_cdc_gov.R
\name{data_cdc_gov_endpoint}
\alias{data_cdc_gov_endpoint}
\title{Construct data.cdc.gov API endpoints from dataset IDs}
\usage{
data_cdc_gov_endpoint(dataset_id)
}
\arguments{
\item{dataset_id}{One or more dataset IDs, as a vector of strings.}
}
\value{
The full URL(s) for the (JSON) Socrata Open Data API
endpoint(s).
}
\description{
Construct data.cdc.gov API endpoints from dataset IDs
}
\examples{

data_cdc_gov_endpoint(data_cdc_gov_dataset_id("nhsn_hrd_prelim"))

data_cdc_gov_dataset_id("nssp_prop_ed_visits") |>
  data_cdc_gov_endpoint()

}
