% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{data_cdc_gov_ids}
\alias{data_cdc_gov_ids}
\title{Dataset IDs for\code{data.cdc.gov} datasets of interest:}
\format{
An object of class \code{list} of length 4.
}
\usage{
data_cdc_gov_ids
}
\description{
\code{nhsn_hist_daily}: National Healthcare Safety Network
(NHSN) daily-resolution incident hospital admissions timeseries,
for COVID-19 and later Influenza, reported until 28 June 2024.
Former target for the COVID-19 Forecast Hub.
}
\details{
\code{nhsn_hrd_final}, \code{nhsn_hrd_prelim}: NHSN Hospital Respiratory Data
(HRD).
Epiweekly-resolution incident hospital admissions
timeseries for COVID-19, Influenza, and RSV.
\code{nhsn_hrd_final} points to the most recent final (Friday) weekly data
release. \code{nhsn_hrd_prelim} points the most recent preliminary (Wednesday)
weekly release.

\code{nssp_prop_ed_visits}: National Syndromic Surveillance
Program (NSSP) data on proportion of incident emergency
deparment (ED) visits due to COVID-19, Influenza, and RSV.
}
\keyword{datasets}
