% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_data_cdc_gov.R
\name{data_cdc_gov_soda_query}
\alias{data_cdc_gov_soda_query}
\title{Helper function for common Socrata open data
API (SODA) queries targeted at \code{data.cdc.gov} datasets.}
\usage{
data_cdc_gov_soda_query(
  dataset_name,
  start_date = NULL,
  end_date = NULL,
  columns = NULL,
  locations = NULL,
  limit = 1e+05,
  order_by = NULL,
  desc = FALSE,
  ...
)
}
\arguments{
\item{dataset_name}{Name of the dataset to query.}

\item{start_date}{Pull only rows with dates
greater than or equal to this date. If \code{NULL},
no minimum date. Default \code{NULL}.}

\item{end_date}{Pull only rows with dates
less than or equal to this date. If \code{NULL},
no maximum date. Default \code{NULL}.}

\item{columns}{Vector of columns to retrieve, in
addition to the dataset's date and location columns, which
are always retrieved. If \code{NULL}, retrieve all columns.
Default \code{NULL}.}

\item{locations}{Vector of locations to
retrieve, in the format of the dataset's location
column. If \code{NULL}, retrieve all. Default \code{NULL}.}

\item{limit}{limit to the number of rows to retrieve.
Default 1e5.}

\item{order_by}{Vector of columns by which to order the
results, if any. Default \code{NULL} (do not order the dataset.}

\item{desc}{whether to order descending instead of
ascending. Default \code{FALSE} (order ascending).}

\item{...}{additional arguments (ignored for now)}
}
\value{
the query as \code{\link[soql:soql]{soql::soql()}} output
}
\description{
Helper function for common Socrata open data
API (SODA) queries targeted at \code{data.cdc.gov} datasets.
}
