% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{default_prism_bin_names}
\alias{default_prism_bin_names}
\title{Names of PRISM bins}
\format{
An object of class \code{character} of length 5.
}
\usage{
default_prism_bin_names
}
\description{
Names of PRISM bins
}
\keyword{datasets}
