% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{default_prism_units}
\alias{default_prism_units}
\title{Default threshold unit for each PRISM surveillance
signal.}
\format{
An object of class \code{character} of length 2.
}
\usage{
default_prism_units
}
\description{
Default threshold unit for each PRISM surveillance
signal.
}
\keyword{datasets}
