% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soda.R
\name{do_soda_query}
\alias{do_soda_query}
\title{Perform an abitrary SOQL query against an arbitrary
dataset, and return the results as a tibble.}
\usage{
do_soda_query(
  query,
  api_key_id = NULL,
  api_key_secret = NULL,
  limit = 1e+05,
  error_on_limit = TRUE,
  api_key_creation_url = NULL
)
}
\arguments{
\item{query}{Full URL representation of the query to perform,
as a \code{\link[soql:soql]{soql::soql()}} query object or a character string.}

\item{api_key_id}{API key id to use when performing the query.
Passed to \code{\link[=.do_api_request]{.do_api_request()}}. Default \code{NULL}.}

\item{api_key_secret}{API key secret to use when performing the query.
Passed to \code{\link[=.do_api_request]{.do_api_request()}}. Default \code{NULL}.
#' @param limit Limit number of rows for the SODA query.}

\item{limit}{Limit number of rows for the SODA query. Default \code{1e5}.}

\item{error_on_limit}{Raise an error if the number
of rows returned is equal to the maximum? Default \code{TRUE}.}

\item{api_key_creation_url}{The user will be warned if they fail
to provide an \code{api_key_id} and \code{api_key_secret}. Provide this
optional argument to direct the user to a website where they can
create those credentials. See \code{\link[=.warn_no_api_creds]{.warn_no_api_creds()}}.}
}
\description{
Perform an abitrary SOQL query against an arbitrary
dataset, and return the results as a tibble.
}
