% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_data_cdc_gov.R
\name{.perform_api_request}
\alias{.perform_api_request}
\title{Default API requests for forecasttools via
httr2}
\usage{
.perform_api_request(
  url,
  api_key_id,
  api_key_secret,
  api_key_creation_url = NULL
)
}
\arguments{
\item{url}{URL for the request to perform, passed
as the url argument to \code{\link[httr2:request]{httr2::request()}}.}

\item{api_key_id}{API key id to use when authenticating.
If \code{NULL} or the empty string (\code{""}), treated as not provided.}

\item{api_key_secret}{API key secret to use when authenticating.
If \code{NULL} or the empty string (\code{""}), treated as not provided.}

\item{api_key_creation_url}{The user will be warned if they fail
to provide an \code{api_key_id} and \code{api_key_secret}. Provide this optional
argument to direct the user to a website where they can create
those credentials. See \code{\link[=.warn_no_api_creds]{.warn_no_api_creds()}}.}
}
\description{
Default API requests for forecasttools via
httr2
}
\keyword{internal}
