% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_data_cdc_gov.R
\name{.process_data_cdc_gov_response}
\alias{.process_data_cdc_gov_response}
\title{Default processing for a SODA response from \code{data.cdc.gov}}
\usage{
.process_data_cdc_gov_response(response, limit, error_on_limit = TRUE)
}
\arguments{
\item{response}{Response to process, as the output of
\code{\link[httr2:req_perform]{httr2::req_perform()}}.}

\item{limit}{Limit number of rows for the SODA query.}

\item{error_on_limit}{Raise an error if the number
of rows returned is equal to the maximum? Default \code{TRUE}.}
}
\description{
Returns a data frame and optionally errors if a query row
limit was hit.
}
\keyword{internal}
