% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soda.R
\name{.process_soda_response}
\alias{.process_soda_response}
\title{Helper functions for working with the
Socrata Open Data API (SODA)
Default processing for a SODA response.}
\usage{
.process_soda_response(response, limit, error_on_limit = TRUE)
}
\arguments{
\item{response}{Response to process, as the output of
\code{\link[httr2:req_perform]{httr2::req_perform()}}.}

\item{limit}{Limit number of rows for the SODA query.}

\item{error_on_limit}{Raise an error if the number
of rows returned is equal to the maximum? Default \code{TRUE}.}
}
\description{
Returns a data frame and optionally errors if a query row
limit was hit.
}
\keyword{internal}
