% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorize_prism.R
\name{.raise_prism_cutpoint_retrieval_error}
\alias{.raise_prism_cutpoint_retrieval_error}
\title{Helper function for raising informative errors
when \code{\link[=get_prism_cutpoints]{get_prism_cutpoints()}} fails.}
\usage{
.raise_prism_cutpoint_retrieval_error(matches, candidates, desired_cutpoints)
}
\description{
Helper function for raising informative errors
when \code{\link[=get_prism_cutpoints]{get_prism_cutpoints()}} fails.
}
