% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hub_utils.R
\name{download_hub}
\alias{download_hub}
\title{Download a forecast hub from \code{hub_url} to directory \code{hub_path}.}
\usage{
download_hub(hub_url, hub_path, force = FALSE)
}
\arguments{
\item{hub_url}{URL to forecast hub on github.}

\item{hub_path}{Path to where forecast hub should be downloaded to.}

\item{force}{If TRUE then if \code{hub_path} directory already exists,
first delete it and then download the forecast hub.}
}
\value{
Absolute path to downloaded hub.
}
\description{
Download a forecast hub from \code{hub_url} to directory \code{hub_path}.
}
