% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncdf_to_dt.R
\name{dt_to_netcdf}
\alias{dt_to_netcdf}
\title{function for writing netcdfs from long data tables.}
\usage{
dt_to_netcdf(
  dt,
  vars,
  units = NULL,
  dim_vars = intersect(c("lon", "lat", "time"), names(dt)),
  dim_var_units = NULL,
  nc_out
)
}
\arguments{
\item{dt}{a data.table}

\item{vars}{names of columns in dt containing variables.}

\item{units}{character vector containing the units for vars (in the same order). If this is NULL (default), the user is prompted for input.}

\item{dim_vars}{names of columns in dt containing dimension variables.}

\item{dim_var_units}{character vector containing the units for dim_vars (in the same order). If this is NULL (default), the user is prompted for input, except for lon/lat, which is filled in automatically.}

\item{nc_out}{(path and) file name of the netcdf to write}
}
\value{
none.
}
\description{
currently does not support writing netcdfs with multiple variables that have different dimension variables!
}
\author{
Claudio
}
