% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiweek.R
\name{epiyear_first_date}
\alias{epiyear_first_date}
\title{Get the first date of an
epidemiological year ("epiyear")}
\usage{
epiyear_first_date(epiyear, epiweek_standard = "MMWR")
}
\arguments{
\item{epiyear}{The epidemiological
year for which to get the first date.}

\item{epiweek_standard}{One of \code{"USA"} or \code{"MMWR"}
(USA / MMWR epiweek, starts on Sunday) and \code{"ISO"} (ISO
week, starts on Monday). Not case-sensitive.
Default \code{"MMWR"}.}
}
\value{
The first date of the epiyear,
as a \code{\link[lubridate:date]{lubridate::date}} object
}
\description{
For the ISO, the epidemiological year ('epiyear') YYYY
begins on the Monday of the Monday to Sunday week
that includes January 4.
}
\details{
The US CDC / MMWR epiyear is analogous
but with a Sunday to Saturday week. That is,
it begins on the Sunday of the Sunday to
Saturday week that includes January 4.

Both of these definitions ensure that an epiyear
always starts on the first day of a seven day
week (Monday for ISO, Sunday for US CDC/MMWR),
and specifically on whichever first day of the
week falls nearest to 1 January.
}
