% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiweek_to_date.R
\name{epiyear_n_weeks}
\alias{epiyear_n_weeks}
\title{Get the number of epidemiological weeks in a given
epidemiological year.}
\usage{
epiyear_n_weeks(epiyear, epiweek_standard = "MMWR")
}
\arguments{
\item{epiyear}{Integer vector of epidemiological year values.}

\item{epiweek_standard}{One of \code{"USA"} or \code{"MMWR"}
(USA / MMWR epiweek, starts on Sunday) and \code{"ISO"} (ISO
week, starts on Monday). Not case-sensitive.
Default \code{"MMWR"}.}
}
\value{
Integer vector of the number(s) of weeks in the given
epidemiological year(s).
}
\description{
Inspired by the \href{https://epiweeks.readthedocs.io/en/stable/_modules/epiweeks.html#Year.totalweeks}{implementation}
in the \href{https://epiweeks.readthedocs.io/}{\code{epiweeks}} Python package.
}
\examples{

epiyear_n_weeks(2023)
epiyear_n_weeks(2025)
epiyear_n_weeks(2026)
epiyear_n_weeks(2026, epiweek_standard = "ISO")
}
