% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ex_inferencedata_dataframe}
\alias{ex_inferencedata_dataframe}
\title{Example InferenceData DataFrame}
\format{
\subsection{\code{ex_inferencedata_dataframe}}{

A data frame with 40 rows and 20 columns:
}
}
\source{
<data-raw/ex_inferencedata_dataframe.R>
}
\usage{
ex_inferencedata_dataframe
}
\description{
An InferenceData Dataframe with
}
\keyword{datasets}
