% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_daily_forecast_flu}
\alias{example_daily_forecast_flu}
\title{An example daily-resolution set of forecast trajectories for influenza.}
\format{
A tibble with 159,000 rows and 4 columns:
\describe{
\item{.draw}{Unique identifier for an individual MCMC draw.}
\item{date}{Date for a forecasted count.}
\item{hosp}{Forecasted count of hospital admissions.}
\item{location}{Location for a forecasted count.}
}
}
\source{
data-raw/example_daily_forecast_flu.R
}
\usage{
example_daily_forecast_flu
}
\description{
53 US states and territories are represented, coded according to their
USPS two-letter abbreviations: the 50 US states, Puerto Rico,
the District of Columbia, and the US Virgin Islands. 100 randomly
chosen draws from an original MCMC sample of 4000 are provided
}
\keyword{datasets}
