% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_for_scoring.R
\name{filter_for_scoring}
\alias{filter_for_scoring}
\title{Filter a table of forecast and target data
intended for scoring.}
\usage{
filter_for_scoring(
  forecasts_and_targets,
  min_forecast_date = NULL,
  max_forecast_date = NULL,
  horizons = NULL,
  locations = NULL,
  models = NULL
)
}
\arguments{
\item{forecasts_and_targets}{table of forecasts
and targets to filter for scoring.}

\item{min_forecast_date}{Only score forecasts dated
on or after this date. If \code{NULL}, no minimum date.
Default \code{NULL}.}

\item{max_forecast_date}{Only score forecasts dated
on or before this date. If \code{NULL}, no maximum date.
Default \code{NULL}.}

\item{horizons}{Only score forecasts with forecast
horizons in this vector. If \code{NULL}, score all horizons
in the table. Default \code{NULL}.}

\item{locations}{Only score forecasts for locations
in this vector of locations. If \code{NULL}, score all
available locations. Default \code{NULL}.}

\item{models}{Only score forecasts for models in
this vector of models. If \code{NULL}, score all
available models. Default \code{NULL}.}
}
\value{
The filtered table
}
\description{
Takes in a table generated by \code{\link[=create_table_for_scoring]{create_table_for_scoring()}}
or \code{\link[scoringutils:score]{scoringutils::score()}} and and filters it according to
various parameters.
}
