% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{filter_largest_lte}
\alias{filter_largest_lte}
\title{Filter a data frame by a column to rows with
the largest value of that column that is
less than or equal to specified maximum value.}
\usage{
filter_largest_lte(df, column, max_value, .by = NULL, .preserve = FALSE)
}
\arguments{
\item{df}{data frame to filter}

\item{column}{column to filter on.}

\item{max_value}{Maximum value. Filter to the
largest value in \code{column} less than or equal to \code{max_value}.}

\item{.by}{Optional grouping columns in \code{df} for the
filter. Passed as the \code{.by} argument to \code{\link[dplyr:filter]{dplyr::filter()}}.
Default \code{NULL}, matching the \code{\link[dplyr:filter]{dplyr::filter()}} default.}

\item{.preserve}{Preserve all groups present in grouped input?
Passed as the \code{.preverse} argument to \code{\link[dplyr:filter]{dplyr::filter()}}.
Default \code{FALSE}, matching the \code{\link[dplyr:filter]{dplyr::filter()}} default.}
}
\value{
The filtered data frame.
}
\description{
Useful for getting the last date prior to or on
a given target date (e.g. matching data vintages).
}
\details{
Uses \code{\link[dplyr:filter]{dplyr::filter()}} syntax, and accepts data-masked
expressions for \code{column} and \code{max_value}.

Returns a 0-row tibble if no rows match the criteria.
}
\examples{
some_dates <- tibble::tibble(
   row_no = 1:3,
   date = as.Date(c("2026-01-01", "2026-07-02", "2026-07-03"))
)

some_dates |> filter_largest_lte(date, as.Date("2026-07-02"))
some_dates |> filter_largest_lte(date, as.Date("2026-07-03"))
some_dates |> filter_largest_lte(date, as.Date("2026-07-01"))

}
