% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_to_shared_forecasts.R
\name{filter_to_shared_forecasts}
\alias{filter_to_shared_forecasts}
\title{Filter a table of forecasts to include only those present
for all values of a given column.}
\usage{
filter_to_shared_forecasts(tbl, comparator_values = NULL, compare = "model")
}
\arguments{
\item{tbl}{Table of forecasts or scores to filter,
as a valid input to \code{\link[scoringutils:get_forecast_unit]{scoringutils::get_forecast_unit()}}.}

\item{comparator_values}{Character vector of comparator values
for which to compute shared forecasts. If \code{NULL} (default), use all
unique values of the column specified in \code{compare}.}

\item{compare}{Name of the column containing the comparator values.
Default \code{"model"}.}
}
\value{
Table filtered to containing only forecasts corresponding to
entries in \code{comparator_values} that are present for \emph{all} entries
in \code{comparator_values}.
}
\description{
Filter a table of forecasts to include only those present
for all values of a given column.
}
\examples{

filter_to_shared_forecasts(
    scoringutils::example_quantile,
    c("EuroCOVIDhub-ensemble", "UMass-MechBayes"))

scoringutils::example_quantile |>
    scoringutils::score() |>
    filter_to_shared_forecasts(c("EuroCOVIDhub-ensemble",
                                 "epiforecasts-EpiNow2",
                                 "UMass-MechBayes"))
}
\seealso{
\code{\link[=get_shared_forecasts]{get_shared_forecasts()}}
}
