% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_to_subset_forecasts.R
\name{filter_to_subset_forecasts}
\alias{filter_to_subset_forecasts}
\title{Filter a table of forecasts to include only those present
for a value of a given column.}
\usage{
filter_to_subset_forecasts(tbl, comparator_value, compare = "model")
}
\arguments{
\item{tbl}{Table of forecasts or scores to filter,
as a valid input to \code{\link[scoringutils:get_forecast_unit]{scoringutils::get_forecast_unit()}}.}

\item{comparator_value}{Character scalar of comparator value
for which to compute subset forecasts.}

\item{compare}{Name of the column containing the comparator value.
Default \code{"model"}.}
}
\value{
Table filtered to containing only forecasts that are also
present in the \code{comparator_value} forecasts.
}
\description{
Filter a table of forecasts to include only those present
for a value of a given column.
}
\examples{

filter_to_subset_forecasts(
    scoringutils::example_quantile,
    "EuroCOVIDhub-ensemble")

scoringutils::example_quantile |>
    scoringutils::score() |>
    filter_to_subset_forecasts("EuroCOVIDhub-ensemble")

}
\seealso{
\code{\link[=get_subset_forecasts]{get_subset_forecasts()}}
}
