% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round_epiweek.R
\name{floor_mmwr_epiweek}
\alias{floor_mmwr_epiweek}
\title{Get the first date in a US/MMWR epidemiological week}
\usage{
floor_mmwr_epiweek(date)
}
\arguments{
\item{date}{A date.}
}
\value{
The date that ends the epidemiological week, as a
\code{\link[lubridate:date]{lubridate::date()}}.
}
\description{
Given any date from a US/MMWR epidemiological week ("epiweek") return
the date of the first day of the epiweek (a Monday).
}
\examples{
floor_mmwr_epiweek(as.Date("2025-03-13")) # rounded down to Sunday
floor_mmwr_epiweek(as.Date("2025-03-08")) # this is already Sunday
}
