% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecasttools-package.R
\docType{package}
\name{forecasttools-package}
\alias{forecasttools}
\alias{forecasttools-package}
\title{forecasttools: R Toolbox for Short-Term Infectious Disease Forecasting}
\description{
A collection of tools for short-term forecasting within CFA Predict.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://cdcgov.github.io/forecasttools/}
  \item \url{https://github.com/CDCgov/forecasttools/}
  \item Report bugs at \url{https://github.com/CDCgov/forecasttools/issues}
}

}
\author{
\strong{Maintainer}: Dylan Morris \email{dzl1@cdc.gov}

Authors:
\itemize{
  \item Dylan Morris \email{dzl1@cdc.gov}
  \item Samuel Brand \email{usi1@cdc.gov}
  \item Damon Bayer \email{xum8@cdc.gov}
}

Other contributors:
\itemize{
  \item Subekshya Bidari \email{zib2@cdc.gov} [contributor]
  \item Adam Howes \email{xwg3@cdc.gov} [contributor]
  \item Kaitlyn Johnson \email{uox1@cdc.gov} [contributor]
}

}
\keyword{internal}
