% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_table_for_scoring.R
\name{gather_hub_forecast_data}
\alias{gather_hub_forecast_data}
\title{Collect and reformat hub forecasts for scoring.}
\usage{
gather_hub_forecast_data(hub_path)
}
\arguments{
\item{hub_path}{Local path to hub}
}
\value{
Forecast data as a table.
}
\description{
Gather hub forecasts with stripped out \code{horizon == -1}
and rename variables in a way to match \code{scoringutils}
expectations.
}
