% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hub_utils.R
\name{gather_hub_location_data}
\alias{gather_hub_location_data}
\title{Gather location data from a forecast hub.}
\usage{
gather_hub_location_data(
  hub_path,
  location_file = fs::path("auxiliary-data", "locations.csv")
)
}
\arguments{
\item{hub_path}{Local path to forecast hub.}

\item{location_file}{The path to the data file for
location data relative to forecast hub directory.
Defaults to expected path in the FluSight forecast hub.}
}
\value{
Table of location data.
}
\description{
Gather location data from a forecast hub.
}
