% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hub_utils.R
\name{gather_hub_quantile_forecasts}
\alias{gather_hub_quantile_forecasts}
\title{Collect and reformat hub forecasts for scoring.}
\usage{
gather_hub_quantile_forecasts(hub_path)
}
\arguments{
\item{hub_path}{Local path to hub}
}
\value{
Forecast data as a table.
}
\description{
Gather hub quantile forecasts.
}
