% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_table_for_scoring.R
\name{gather_location_data}
\alias{gather_location_data}
\title{Gather location data from the forecast hub.}
\usage{
gather_location_data(
  hub_path,
  local_datapath = fs::path("auxiliary-data", "locations.csv")
)
}
\arguments{
\item{hub_path}{Local path to forecast hub.}

\item{local_datapath}{The local path to the data file for
location data relative to forecast hub directory.
Defaults to expected path in the FluSight forecast hub.}
}
\value{
Table of location data.
}
\description{
This includes population
size for the states.
}
