% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_table_for_scoring.R
\name{gather_target_data}
\alias{gather_target_data}
\title{Gather target truth data
from the forecast hub and prepare it
for use with \code{scoringutils}.}
\usage{
gather_target_data(
  hub_path,
  local_datapath = fs::path("target-data", "target-hospital-admissions.csv")
)
}
\arguments{
\item{hub_path}{Local path to hub.}

\item{local_datapath}{The local path to the truth data file for
relative to forecast hub directory.
Defaults to expected path in the FluSight forecast hub.}
}
\value{
Table of target truth data
}
\description{
This requires renaming the \code{value} col to
\code{truth_value} to comply with \code{scoringutils}
expectations.
}
