% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_path_point.R
\name{geom_path_point}
\alias{geom_path_point}
\alias{geom_line_point}
\title{Plot a path of connected points}
\usage{
geom_path_point(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  lineend = "butt",
  linejoin = "round",
  linemitre = 10,
  arrow = NULL,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)

geom_line_point(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  na.rm = FALSE,
  orientation = NA,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[ggplot2:aes]{aes()}}. If specified and
\code{inherit.aes = TRUE} (the default), it is combined with the default mapping
at the top level of the plot. You must supply \code{mapping} if there is no plot
mapping.}

\item{data}{The data to be displayed in this layer. There are three
options:

If \code{NULL}, the default, the data is inherited from the plot
data as specified in the call to \code{\link[ggplot2:ggplot]{ggplot()}}.

A \code{data.frame}, or other object, will override the plot
data. All objects will be fortified to produce a data frame. See
\code{\link[ggplot2:fortify]{fortify()}} for which variables will be created.

A \code{function} will be called with a single argument,
the plot data. The return value must be a \code{data.frame}, and
will be used as the layer data. A \code{function} can be created
from a \code{formula} (e.g. \code{~ head(.x, 10)}).}

\item{stat}{The statistical transformation to use on the data for this layer.
When using a \verb{geom_*()} function to construct a layer, the \code{stat}
argument can be used to override the default coupling between geoms and
stats. The \code{stat} argument accepts the following:
\itemize{
\item A \code{Stat} ggproto subclass, for example \code{StatCount}.
\item A string naming the stat. To give the stat as a string, strip the
function name of the \code{stat_} prefix. For example, to use \code{stat_count()},
give the stat as \code{"count"}.
\item For more information and other ways to specify the stat, see the
\link[ggplot2:layer_stats]{layer stat} documentation.
}}

\item{position}{A position adjustment to use on the data for this layer. This
can be used in various ways, including to prevent overplotting and
improving the display. The \code{position} argument accepts the following:
\itemize{
\item The result of calling a position function, such as \code{position_jitter()}.
This method allows for passing extra arguments to the position.
\item A string naming the position adjustment. To give the position as a
string, strip the function name of the \code{position_} prefix. For example,
to use \code{position_jitter()}, give the position as \code{"jitter"}.
\item For more information and other ways to specify the position, see the
\link[ggplot2:layer_positions]{layer position} documentation.
}}

\item{lineend}{Line end style (round, butt, square).}

\item{linejoin}{Line join style (round, mitre, bevel).}

\item{linemitre}{Line mitre limit (number greater than 1).}

\item{arrow}{Arrow specification, as created by \code{\link[grid:arrow]{grid::arrow()}}.}

\item{na.rm}{If \code{FALSE}, the default, missing values are removed with
a warning. If \code{TRUE}, missing values are silently removed.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display. To include legend keys for all levels, even
when no data exists, use \code{TRUE}.  If \code{NA}, all levels are shown in legend,
but unobserved levels are omitted.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[ggplot2:annotation_borders]{annotation_borders()}}.}

\item{...}{Other arguments passed on to \code{\link[ggplot2:layer]{layer()}}'s \code{params} argument. These
arguments broadly fall into one of 4 categories below. Notably, further
arguments to the \code{position} argument, or aesthetics that are required
can \emph{not} be passed through \code{...}. Unknown arguments that are not part
of the 4 categories below are ignored.
\itemize{
\item Static aesthetics that are not mapped to a scale, but are at a fixed
value and apply to the layer as a whole. For example, \code{colour = "red"}
or \code{linewidth = 3}. The geom's documentation has an \strong{Aesthetics}
section that lists the available options. The 'required' aesthetics
cannot be passed on to the \code{params}. Please note that while passing
unmapped aesthetics as vectors is technically possible, the order and
required length is not guaranteed to be parallel to the input data.
\item When constructing a layer using
a \verb{stat_*()} function, the \code{...} argument can be used to pass on
parameters to the \code{geom} part of the layer. An example of this is
\code{stat_density(geom = "area", outline.type = "both")}. The geom's
documentation lists which parameters it can accept.
\item Inversely, when constructing a layer using a
\verb{geom_*()} function, the \code{...} argument can be used to pass on parameters
to the \code{stat} part of the layer. An example of this is
\code{geom_area(stat = "density", adjust = 0.5)}. The stat's documentation
lists which parameters it can accept.
\item The \code{key_glyph} argument of \code{\link[ggplot2:layer]{layer()}} may also be passed on through
\code{...}. This can be one of the functions described as
\link[ggplot2:draw_key]{key glyphs}, to change the display of the layer in the legend.
}}

\item{orientation}{The orientation of the layer. The default (\code{NA})
automatically determines the orientation from the aesthetic mapping. In the
rare event that this fails it can be given explicitly by setting \code{orientation}
to either \code{"x"} or \code{"y"}. See the \emph{Orientation} section for more detail.}
}
\description{
Adding \code{geom_path_point()} or \code{geom_line_point()} to a
\code{\link[ggplot2:ggplot]{ggplot}} is similar to adding
\code{\link[ggplot2:geom_path]{geom_path()}} followed by
\code{\link[ggplot2:geom_point]{geom_point()}} or
\code{\link[ggplot2:geom_line]{geom_line()}} followed by
\code{\link[ggplot2:geom_point]{geom_point()}}, respectively.
The difference is that that with the unified \code{geom_path_point()}
and \code{geom_line_point()} geoms, paths/lines and points are
plotted together, as a single \code{\link[ggplot2:layer_geoms]{layer}}.
In the stacked approach, they are plotted as two separate layers.
See Details for a discussion of why this can matter in practice.
}
\details{
Plotting with a single layer is important when the user wants
points and lines to be stacked ("z-ordered") \emph{together} by
\link[ggplot2:group]{group}, i.e. all points and lines representing
group 1 to be below all points and lines in group 2, et cetera.
With the standard \code{geom_line() + geom_point()} approach, all
points (z-ordered by group) will be plotted on top of all lines
(z-ordered by group), or the reverse if one does
\code{geom_point() + geom_line()}. this can make the plot more difficult
to read when points partially overlap. See the Examples for a
demonstration.
}
\examples{
library(ggplot2)
library(tibble)
data <- tibble(x = rep(c(3,1,2), 2),
               y = c(-0.5, 0.5, 0, 0.5, 0, 0.02),
               z = c(rep("a", 3), rep("b", 3)))

data <- tribble(
   ~x,      ~y,   ~z,
    1,       0,  "a",
    2,       0,  "a",
    3,    -0.5,  "a",
    1,    0.01,  "b",
    3,   -0.520,  "b",
    2,     0.02,  "b")

# standard approach causes points to "interrupt"
# lines when overplotted
ggplot(data, aes(x = x, y = y, color = z)) +
   geom_path(linewidth = 2) +
   geom_point(size = 5)

# using a single-layer approach fixes this
ggplot(data, aes(x = x, y = y, color = z)) +
   geom_path_point(linewidth = 2, size = 5)

# The same is true for `geom_line` and `geom_line_point`
# multi-layer approach:
ggplot(data, aes(x = x, y = y, color = z)) +
   geom_line(linewidth = 3) +
   geom_point(size = 5)

# single-layer approach:
ggplot(data, aes(x = x, y = y, color = z)) +
   geom_line_point(linewidth = 3, size = 5)

}
