% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_path_point.R
\name{GeomPathPoint}
\alias{GeomPathPoint}
\alias{GeomLinePoint}
\title{Geometric object showing a path of connected points.}
\usage{
GeomPathPoint

GeomLinePoint
}
\description{
Geometric object showing a path of connected points.

Geometric object showing a path of connected points sorted by
orderded in the order of the x-axis variable.
}
