% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hubverse_quantiles_to_median_qi.R
\name{get_available_qi}
\alias{get_available_qi}
\title{Get a table of quantile intervals ("qi") that can be computed
from a given set of absolute quantile levels.}
\usage{
get_available_qi(quantile_levels, quantile_tol = 10)
}
\arguments{
\item{quantile_levels}{Vector of quantile levels from which
to get available quantile intervals.}

\item{quantile_tol}{Round quantile level values to this many
decimal places, to avoid problems with floating point number
equality comparisons. Passed as the \code{digits} argument to
\code{\link[base:Round]{base::round()}}. Default 10.}
}
\value{
A table of available intervals, as a
\code{\link[tibble:tibble]{tibble}}, with columns
\code{.lower_quantile}, \code{.upper_quantile}, and \code{.width},
that give the bounds and the widths of the available
intervals.
}
\description{
These intervals are also called "equal-tailed intervals". A quantile
interval can be computed whenver both the \code{x} quantile and the
\code{1 - x} quantile are present, yielding an interval of width
\code{abs(1 - 2 * x)}.
}
\details{
Checks that all quantile_levels are between 0 and 1, inclusive.
}
\examples{
get_available_qi(c(0.25, 0.3, 0.5, 0.75, 0.99))
get_available_qi(c(0.1, 0.2, 0.5, 0.8, 0.9))
}
