% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiweek_to_date.R
\name{get_epiweek_start}
\alias{get_epiweek_start}
\title{Get the start day of an epiweek.}
\usage{
get_epiweek_start(epiweek_standard)
}
\arguments{
\item{epiweek_standard}{One of \code{"MMWR"} or \code{"USA"}
(USA / MMWR epiweek, starts on Sunday) and \code{"ISO"} (ISO
week, starts on Monday). Not case-sensitive.}
}
\value{
the start day of the week for the
epiweek standard, as an integer that can be passed to
\code{\link[lubridate:day]{lubridate::wday()}} as the \code{week_start} argument.
}
\description{
Get the start day of an epiweek
according to different epiweek
standards. Returns it as a string
that can be passed to \code{\link[lubridate:day]{lubridate::wday()}}
as the \code{week_start} argument.
}
