% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_hubverse.R
\name{get_hubverse_table}
\alias{get_hubverse_table}
\title{Given a set of epiweekly quantile forecasts,
format them for a hubverse submission}
\usage{
get_hubverse_table(
  quantile_forecasts,
  reference_date,
  target_name,
  quantile_value_col = "quantile_value",
  quantile_level_col = "quantile_level",
  location_col = "location",
  epiweek_col = "epiweek",
  epiyear_col = "epiyear",
  horizons = -1:3,
  horizon_timescale = c("days", "weeks"),
  reference_dow = 7,
  week_start = 7,
  excluded_locations = c("60", "78")
)
}
\arguments{
\item{quantile_forecasts}{tidy data frame of
quantile forecasts to format,
which should have a location column, a column of
quantile values, a column of quantile levels,
a column indicating the epiweek, and a column
indicating the epiyear.}

\item{reference_date}{reference date for the hub forecast.}

\item{target_name}{Name of the target for the table.}

\item{quantile_value_col}{name of the column containing the quantile
values. Default \code{'quantile_value'} (the default output
column name from \code{\link[=trajectories_to_quantiles]{trajectories_to_quantiles()}}).}

\item{quantile_level_col}{name of the column indicating the
quantile level. Default \code{'quantile_level'} (the default output
column name from \code{\link[=trajectories_to_quantiles]{trajectories_to_quantiles()}}).}

\item{location_col}{name of the column containing the location
value. Default \code{'location'}.}

\item{epiweek_col}{name of the column containing the epiweek.
Default \code{'epiweek'}.}

\item{epiyear_col}{name of the column containing the epiyear.
Default \code{'epiyear'}.}

\item{horizons}{Vector of forecast horizons to include,
in weeks ahead of the \code{reference_date}.
Default -1:3 (FluSight and Covidhub 2024/25 horizons).}

\item{horizon_timescale}{Either "days" or "weeks"}

\item{reference_dow}{Which day of the week should the reference_date
be, as an integer? Default 7 (the last day of the week), which
with the default value of \code{week_start} will be Saturday, i.e.
the last day of a USA epidemiological week.}

\item{week_start}{Starting day of the week relative to
Monday, as an integer. Default 7, i.e. weeks that start on
Sunday, as USA epidemiological weeks do. See the \code{\link[lubridate:day]{lubridate::wday()}}
documentation for details.}

\item{excluded_locations}{locations to exclude from the table
Default c("60", "78") (American Samoa and the US Virgin Islands,
for which FluSight does not currently accept forecasts).}
}
\value{
a properly formatted table, as a \code{\link[tibble:tibble]{tibble}}.
}
\description{
Given a set of epiweekly quantile forecasts,
format them for a hubverse submission
}
