% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_to_shared_forecasts.R
\name{get_shared_forecasts}
\alias{get_shared_forecasts}
\title{Compute the set of forecasts present for all values of a given
column.}
\usage{
get_shared_forecasts(tbl, comparator_values = NULL, compare = "model")
}
\arguments{
\item{tbl}{Table of forecasts or scores to filter,
as a valid input to \code{\link[scoringutils:get_forecast_unit]{scoringutils::get_forecast_unit()}}.}

\item{comparator_values}{Character vector of comparator values
for which to compute shared forecasts. If \code{NULL} (default), use all
unique values of the column specified in \code{compare}.}

\item{compare}{Name of the column containing the comparator values.
Default \code{"model"}.}
}
\value{
Table of forecast unit values for all and only the shared
forecasts, which can be joined to the original table for filtering
purposes (see \code{\link[=filter_to_shared_forecasts]{filter_to_shared_forecasts()}}).
}
\description{
Compute the set of forecasts present for all values of a given
column.
}
\examples{

get_shared_forecasts(scoringutils::example_quantile,
                           c("EuroCOVIDhub-ensemble", "UMass-MechBayes"))

}
\seealso{
\code{\link[=filter_to_shared_forecasts]{filter_to_shared_forecasts()}}
}
