% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_to_subset_forecasts.R
\name{get_subset_forecasts}
\alias{get_subset_forecasts}
\title{Compute the set of forecasts present for a given value of a given
column.}
\usage{
get_subset_forecasts(tbl, comparator_value, compare = "model")
}
\arguments{
\item{tbl}{Table of forecasts or scores to filter,
as a valid input to \code{\link[scoringutils:get_forecast_unit]{scoringutils::get_forecast_unit()}}.}

\item{comparator_value}{Character scalar of comparator values
for which to compute shared forecasts.}

\item{compare}{Name of the column containing the comparator values.
Default \code{"model"}.}
}
\value{
Table of forecast unit values for all and only the subset
forecasts, which can be joined to the original table for filtering
purposes (see \code{\link[=filter_to_subset_forecasts]{filter_to_subset_forecasts()}}).
}
\description{
Compute the set of forecasts present for a given value of a given
column.
}
\examples{

get_subset_forecasts(scoringutils::example_quantile, "EuroCOVIDhub-ensemble")

}
\seealso{
\code{\link[=filter_to_subset_forecasts]{filter_to_subset_forecasts()}}
}
