% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms.R
\name{get_transform_to_real_line}
\alias{get_transform_to_real_line}
\title{Get a monotonic transform from a real interval
to the unconstrained real line.}
\usage{
get_transform_to_real_line(lb, ub)
}
\arguments{
\item{lb}{Lower bound of the interval. \code{-Inf} for
no lower bound.}

\item{ub}{Upper bound of the interval. \code{Inf} for
no upper bound.}
}
\value{
The transform, as a callable object.
}
\description{
Uses \code{\link[=scaled_shifted_log]{scaled_shifted_log()}} for intervals of the form
\eqn{(-\infty, x]} or \eqn{[x, \infty]} with finite \eqn{x}.
Uses \code{\link[=scaled_logit]{scaled_logit()}} for intervals of
the form \eqn{(x, y)} with finite \eqn{x, y}. Uses the identity
transform (\code{\link[base:identity]{base::identity()}}) for the real line
\eqn{(\-infty, infty)}.
}
\examples{

my_scaled_logit <- get_transform_to_real_line(-100, 100)
my_scaled_logit(0)
my_scaled_logit(50)

my_shifted_log <- get_transform_to_real_line(6, Inf)
my_shifted_log(20)

my_other_shifted_log <- get_transform_to_real_line(-Inf, 10)
my_other_shifted_log(-70)

# returns the identity transform if given no bounds
my_transform <- get_transform_to_real_line(-Inf, Inf)
my_transform(50)

}
