% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{ggplot_dt}
\alias{ggplot_dt}
\title{plotting function for precip fields}
\usage{
ggplot_dt(
  dt,
  data_col = colnames(dt)[3],
  rr = NULL,
  mn = NULL,
  colorscale = scale_fill_gradient2(low = "blue", mid = "white", high = "red", name =
    data_col, limits = rr),
  mapscale = "large"
)
}
\arguments{
\item{dt}{Data table containing columns lon, lat and the data. The data defaults to the third column. If multiple time-steps are available in dt, only the the data to the first one is plotted.}

\item{data_col}{The name of the column in dt containing the data.}

\item{rr}{range of the plotted variable, default is the range of the data. Manipulate this if you compare multiple plots.}

\item{mn}{plot title}

\item{colorscale}{A ScaleContinuous object specifying the scale for the plot layer fill. Can be used to change colors or legend title (by setting name).}

\item{mapscale}{Either 'low', 'medium' or 'high'. Controls the scale of the map. Higher scale leads to better resolved}

\item{legend_title}{legend title.}
}
\value{
a ggplot object.
}
\description{
plots dat on a regular grid over a map.
}
\author{
Claudio Heinrich
}
