% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_hubverse.R
\name{horizons_from_target_end_dates}
\alias{horizons_from_target_end_dates}
\title{Compute Horizon from Target End Dates}
\usage{
horizons_from_target_end_dates(
  reference_date,
  target_end_dates,
  horizon_timescale
)
}
\arguments{
\item{reference_date}{hub forecasting reference date}

\item{target_end_dates}{vector of target end dates}

\item{horizon_timescale}{Either "days" or "weeks"}
}
\value{
a vector of forecast horizons, in the specified timescale
}
\description{
Compute Horizon from Target End Dates
}
