% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hubverse_to_scorable.R
\docType{data}
\name{hub_output_type_to_scoringutils}
\alias{hub_output_type_to_scoringutils}
\title{Map of hubverse output types to associated
scoringutils scorable forecast types.}
\format{
An object of class \code{character} of length 6.
}
\usage{
hub_output_type_to_scoringutils
}
\description{
Map of hubverse output types to associated
scoringutils scorable forecast types.
}
\keyword{datasets}
