% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_quantiles_to_median_qi.R
\name{hub_quantiles_to_median_qi}
\alias{hub_quantiles_to_median_qi}
\title{Convert a hubverse-format table of quantiles into a
\code{ggdist}-format table of quantile intervals.}
\usage{
hub_quantiles_to_median_qi(
  hubverse_quantile_table,
  .width = NULL,
  require_only_quantiles = TRUE,
  require_all_medians = TRUE,
  require_all_widths = TRUE,
  quantile_tol = 10
)
}
\arguments{
\item{hubverse_quantile_table}{Hubverse-format quantile table,
as an object something coerceible to a \code{\link[tibble:tibble]{tibble}}.}

\item{.width}{Vector of interval width(s) to for which to create rows
in the output. If \code{NULL} (the default), use all and only the widths
that can be computed from quantiles available in
\code{hubverse_quantile_table}, as determined by \code{\link[=get_available_qi_widths]{get_available_qi_widths()}}.}

\item{require_only_quantiles}{Boolean. Error if the table contains
\code{output_type} values other than \code{"quantile"}? If \code{FALSE}, the function
will silently filter the table to rows with \code{output_type == 'quantile'}.
Default \code{TRUE}.}

\item{require_all_medians}{Boolean. Error if the output is missing
any medians? If \code{TRUE}, the function will succeed only if it can find a
median (\code{0.5} quantile) value for all groups present in the input
and otherwise will error. Default \code{TRUE}.}

\item{require_all_widths}{Boolean. Error if the output is missing
any of the requested \code{.width}s for any group? If \code{TRUE}, the
function will succeed only if it can calculate all the requested
interval \code{.width}s for all the groups present in the input and
otherwise will error. Default \code{TRUE}.}

\item{quantile_tol}{Round values in the \code{quantile_level_col} to
this many decimal places, to avoid problems with floating point number
equality comparisons. Passed as the \code{digits} argument to
\code{\link[base:round]{base::round()}}. Default 10.}
}
\value{
A \code{\link[tibble:tibble]{tibble}} in the output format of
\code{\link[ggdist:median_qi]{ggdist::median_qi()}}.
}
\description{
Expects a hubverse-format quantile table (e.g. as created by
\code{\link[=get_hubverse_quantile_table]{get_hubverse_quantile_table()}}). Treats all columns in that table other
than \code{output_type_id}, \code{output_type}, and \code{value} as grouping / id
columns, and returns them in the output table.
}
\examples{
hub_quantiles_to_median_qi(
  hubExamples::forecast_outputs |>
    dplyr::filter(.data$output_type == "quantile"),
  .width = c(0.5, 0.8)
)

hub_quantiles_to_median_qi(hubExamples::forecast_outputs,
  .width = c(0.5),
  require_only_quantiles = FALSE
)

hub_quantiles_to_median_qi(hubExamples::forecast_outputs,
  require_only_quantiles = FALSE,
  require_all_widths = FALSE
)

}
