% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hub_summary_table.R
\name{hub_summary_table}
\alias{hub_summary_table}
\title{Summarize a hubverse format hub into a
table.}
\usage{
hub_summary_table(path_to_hub, reference_dates = NULL, locations = NULL)
}
\description{
Combining one or more forecasts
with model metadata and particular
quantiles.
}
