% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hubverse_to_scorable.R
\name{hub_to_scorable_quantiles}
\alias{hub_to_scorable_quantiles}
\title{Create a table for scoring hub model
quantile forecasts from a local copy of
a hub.}
\usage{
hub_to_scorable_quantiles(hub_path, quantile_tol = 10)
}
\arguments{
\item{hub_path}{Local path to a hubverse-style
forecast hub.}

\item{quantile_tol}{Round quantile level values to this many
decimal places, to avoid problems with floating point number
equality comparisons. Passed as the \code{digits} argument to
\code{\link[base:round]{base::round()}}. Default 10.}
}
\value{
Scorable table, as the output of
\code{\link[scoringutils:as_forecast_quantile]{scoringutils::as_forecast_quantile()}}.
}
\description{
Requires a local version of the
forecast hub at \code{hub_path}.
}
