% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hubverse_to_scoreable.R
\name{hub_to_scoreable_quantiles}
\alias{hub_to_scoreable_quantiles}
\title{Create a table for scoring hub model
quantile forecasts from a local copy of
a hub.}
\usage{
hub_to_scoreable_quantiles(hub_path, ...)
}
\arguments{
\item{hub_path}{Local path to hubverse-style
forecast hub.}

\item{...}{keyword arguments passed to
\code{\link[=hub_quantiles_to_scoreable]{hub_quantiles_to_scoreable()}}.}
}
\value{
Scoreable table, as the output of
\code{\link[scoringutils:as_forecast_quantile]{scoringutils::as_forecast_quantile()}}.
}
\description{
Requires a local version of the
forecast hub at \code{hub_path}.
}
