% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{hubverse_column_names}
\alias{hubverse_column_names}
\title{Internal and user-exposed constants for forecasttools.
Names of the standard columns in a hubverse table.}
\format{
An object of class \code{character} of length 8.
}
\usage{
hubverse_column_names
}
\description{
Internal and user-exposed constants for forecasttools.
Names of the standard columns in a hubverse table.
}
\keyword{datasets}
