% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{hubverse_output_types}
\alias{hubverse_output_types}
\title{Internal and user-exposed constants for forecasttools.
Names of the valid output_type values for hubverse tables.}
\usage{
hubverse_output_types
}
\description{
Internal and user-exposed constants for forecasttools.
Names of the valid output_type values for hubverse tables.
}
