% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{hubverse_std_colnames}
\alias{hubverse_std_colnames}
\title{Names of the standard columns for hubverse tables.
Alias for \link[hubUtils:std_colnames]{hubUtils::std_colnames}.}
\usage{
hubverse_std_colnames
}
\description{
Names of the standard columns for hubverse tables.
Alias for \link[hubUtils:std_colnames]{hubUtils::std_colnames}.
}
