% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{hubverse_valid_output_types}
\alias{hubverse_valid_output_types}
\title{Internal and user-exposed constants for forecasttools.
Names of the valid output_type values for hubverse tables.}
\format{
An object of class \code{character} of length 6.
}
\usage{
hubverse_valid_output_types
}
\description{
Internal and user-exposed constants for forecasttools.
Names of the valid output_type values for hubverse tables.
}
\keyword{datasets}
