% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferencedata_df_to_tidy_draws.R
\name{idata_names_to_tidy_names}
\alias{idata_names_to_tidy_names}
\title{Convert InferenceData column names to tidy column names}
\usage{
idata_names_to_tidy_names(column_names)
}
\arguments{
\item{column_names}{A character vector of InferenceData column names}
}
\value{
A character vector of tidy column names
}
\description{
InferenceData column names for scalar variables are of the form
\code{"('group', 'var_name')"}, while column names for array variables are of the
form \code{"('group', 'var_name[i,j]', 'i_name', 'j_name')"}.
This function converts these column names to a format that is useful for
creating tidy_draws data frames.
\code{"('group', 'var_name')"} becomes \code{"group|var_name"}
\code{"('group', 'var_name[i,j]', 'i_name', 'j_name')"} becomes
\code{"group|var_name[i_name, j_name]"}
}
\examples{
idata_names_to_tidy_names(c(
  "('group', 'var_name')",
  "group|var_name[i_name, j_name]"
))
}
