% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferencedata_df_to_tidy_draws.R
\name{inferencedata_to_tidy_draws}
\alias{inferencedata_to_tidy_draws}
\title{Convert InferenceData DataFrame to nest tidy_draws}
\usage{
inferencedata_to_tidy_draws(idata)
}
\arguments{
\item{idata}{InferenceData DataFrame (the result of calling arviz.InferenceData.to_dataframe in Python)}
}
\value{
A nested tibble, with columns group and data. Each element of data is a tidy_draws data frame
}
\description{
Convert InferenceData DataFrame to nest tidy_draws
}
