% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{isoweek_end}
\alias{isoweek_end}
\title{Integer end date of an ISO week in ISO numbering.
Value (7) corresponds to Sunday.}
\format{
An object of class \code{integer} of length 1.
}
\usage{
isoweek_end
}
\description{
Integer end date of an ISO week in ISO numbering.
Value (7) corresponds to Sunday.
}
\keyword{datasets}
