% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{isoweek_start}
\alias{isoweek_start}
\title{Integer start date of an ISO week in ISO numbering.
Value (1) corresponds to Monday.}
\format{
An object of class \code{integer} of length 1.
}
\usage{
isoweek_start
}
\description{
Integer start date of an ISO week in ISO numbering.
Value (1) corresponds to Monday.
}
\keyword{datasets}
