% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode_locations.R
\name{location_lookup}
\alias{location_lookup}
\title{Look up rows of the USA location table
corresponding to the entries of a given
vector.}
\usage{
location_lookup(location_vector, location_format)
}
\arguments{
\item{location_vector}{vector of location values}

\item{location_format}{format in which the location
vector is coded.
Permitted formats are \code{"abbr"} (state/territory
or nation two letter USPS abbreviation), \code{"hub"}
(legacy 2-digit FIPS code for states and territories, \code{US}
for the USA as a whole), and \code{"long_name"} (full English
name for jurisdiction).}
}
\value{
the rows of the \link{us_location_table}
matching the location vector, with repeats possible.
}
\description{
Get the rows of
\link{us_location_table}
corresponding to a given location vector
and format, with repeats possible
}
