% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{mat_to_dt}
\alias{mat_to_dt}
\title{Turn a matrix with lat/lon idices into a data.table}
\usage{
mat_to_dt(X, lons, lats, var_name = "X")
}
\description{
Turn a matrix with lat/lon idices into a data.table
}
