% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{mmwr_epiweek_end}
\alias{mmwr_epiweek_end}
\title{Integer end date of a USA MMWR epiweek in ISO numbering.
Value (6) corresponds to Saturday.}
\format{
An object of class \code{integer} of length 1.
}
\usage{
mmwr_epiweek_end
}
\description{
Integer end date of a USA MMWR epiweek in ISO numbering.
Value (6) corresponds to Saturday.
}
\keyword{datasets}
