% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{mmwr_epiweek_start}
\alias{mmwr_epiweek_start}
\title{Integer start date of a USA MMWR epiweek in ISO numbering.
Value (7) corresponds to Sunday.}
\format{
An object of class \code{integer} of length 1.
}
\usage{
mmwr_epiweek_start
}
\description{
Integer start date of a USA MMWR epiweek in ISO numbering.
Value (7) corresponds to Sunday.
}
\keyword{datasets}
