% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_nhsn_hrd_admissions.R
\name{nhsn_hrd_admissions_column_names}
\alias{nhsn_hrd_admissions_column_names}
\title{Get the NHSN HRD column names for a disease's weekly
hospital admissions.}
\usage{
nhsn_hrd_admissions_column_names(disease)
}
\arguments{
\item{disease}{Character. One of \code{"covid"}, \code{"flu"},
or \code{"rsv"}.}
}
\value{
Named character vector with \code{count} and \code{rate}
entries giving the corresponding HRD column names.
}
\description{
Get the NHSN HRD column names for a disease's weekly
hospital admissions.
}
\examples{
nhsn_hrd_admissions_column_names("covid")
}
