% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{nullable_comparison}
\alias{nullable_comparison}
\title{Compare two quantities \code{a} and \code{b} with a given
\verb{comparison operator},
returning a single \code{TRUE} if \code{b} is
\code{NULL}.}
\usage{
nullable_comparison(a, comparison_operator, b)
}
\arguments{
\item{a}{First set of values for the comparison}

\item{comparison_operator}{Comparison operator for the comparison,
as a string, e.g. \code{"=="}, \code{">="} \code{"\%in\%"}.}

\item{b}{Second set of values for the comparison, or \code{NULL}.}
}
\value{
A logical vector. Equivalent to \verb{b \{comparison_operator\} a}
if \code{b} is not \code{NULL} and to \code{TRUE} if \code{b} is \code{NULL}.
#'
}
\description{
Useful for letting \code{NULL} reflect "all values"
if \code{\link[dplyr:filter]{dplyr::filter()}} calls. Internal function.
}
\examples{

x <- 6
nullable_comparison(5, ">", x)

x <- NULL
nullable_comparison(5, ">", x)

df <- tibble::tibble(y = 1:6)
x <- 3
df |> dplyr::filter(
  nullable_comparison(y, ">", !!x),
  y < 5
)
x <- NULL
df |> dplyr::filter(
  nullable_comparison(y, ">", !!x),
  y < 5
)

}
