% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivot_hubverse_quantiles_wider.R
\name{pivot_hubverse_quantiles_wider}
\alias{pivot_hubverse_quantiles_wider}
\title{Pivot a hubverse quantile table wider with columns representing
individual quantile levels.}
\usage{
pivot_hubverse_quantiles_wider(
  hubverse_table,
  pivot_quantiles = c(point = 0.5, lower = 0.025, upper = 0.975)
)
}
\arguments{
\item{hubverse_table}{hubverse-format forecast table to pivot,
as a \code{\link[tibble:tibble]{tibble}}}

\item{pivot_quantiles}{quantiles to pivot to columns, as a vector or
named vector. Default \code{c("point" = 0.5, "lower" = 0.025, "upper" = 0.975)},
i.e. get the median and the central 95\% interquantile interval,
with names "point" for the median, "lower" for the 0.025th quantile,
and "upper" for the 0.975th quantile.}
}
\value{
A pivoted version of the hubverse table in which each
forecast for a given target, horizon, reference_date, and
location corresponds to a single row with multiple
value columns, one for each of the quantiles in \code{pivot_quantiles},
and named according to the corresponding names given in that vector,
or generically as \verb{q_<quantile_level>} if an unnamed numeric
vector is provided.
So with the default \code{pivot_quantiles}, the output will have three
value columns named \verb{lower"}, \code{"point"}, and \code{"upper"}
}
\description{
Pivot a hubverse quantile table wider with columns representing
individual quantile levels.
}
