% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_coverage.R
\name{plot_coverage_by_date}
\alias{plot_coverage_by_date}
\title{Plot of empirical forecast coverage by
reference date.}
\usage{
plot_coverage_by_date(
  scored,
  coverage_level,
  coverage_col = NULL,
  date_col = "reference_date",
  group_cols = c("target", "horizon"),
  y_transform = "identity",
  y_labels = scales::label_percent()
)
}
\arguments{
\item{scored}{Output of \code{\link[scoringutils:score]{scoringutils::score()}}, not yet
summarized, containing a column for coverage at the required
coverage level.}

\item{coverage_level}{Decimal coverage level to plot, e.g.
\code{0.95} or \code{0.5}.}

\item{coverage_col}{Name of the column corresponding to that
coverage level in \code{scored}. Default \verb{interval_coverage_<x>}
where \if{html}{\out{<x>}} is the coverage level as a percentage, e.g. if
\code{coverage_level = 0.95}, then if \code{coverage_col} is not specified,
\code{plot_coverage_by_date} will look for a column named
\code{interval_coverage_95}, as this is the default name for
interval coverage columns produced by \code{\link[scoringutils:score]{scoringutils::score()}} and
\code{\link[scoringutils:summarise_scores]{scoringutils::summarise_scores()}}}

\item{date_col}{Column containing dates, which will become the
x-axis in the empirical coverage by date plot. This can be
a target date, but more commonly it will be a forecast date,
or \code{reference_date} indicating when the forecast was produced.
Default \code{"reference_date"}, the standard name for a forecast
date in the hubverse schema.}

\item{group_cols}{Other columns to group by, in addition to forecast
date. These will become facets in the output ggplot. Default
\code{c("target", "horizon")} (i.e. group by forecasting target and
forecast horizon.}

\item{y_transform}{transform for the y axis, a string. Passed
as the \code{transform} argument to \code{\link[ggplot2:scale_y_continuous]{ggplot2::scale_y_continuous()}}.
Default \code{"identity"}.}

\item{y_labels}{labeling scheme for the y axis. Passed as
the \code{labels} argument to \code{\link[ggplot2:scale_y_continuous]{ggplot2::scale_y_continuous()}}. Default
\code{\link[scales:label_percent]{scales::label_percent()}}.}
}
\value{
A ggplot of the empirical coverage.
}
\description{
Plot of empirical forecast coverage by
reference date.
}
