% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_diagnostic}
\alias{plot_diagnostic}
\title{Quick diagnostic plotting function}
\usage{
plot_diagnostic(
  dt,
  var = colnames(dt)[3],
  mn = var,
  rr = NULL,
  col_scheme = "bwr",
  set_white = NULL,
  xlab = "",
  ylab = "",
  save_pdf = FALSE,
  save_dir = "./figures/",
  file_name = "diag_plot",
  stretch_par = NULL,
  legend = TRUE
)
}
\arguments{
\item{dt}{The data table.}

\item{var}{Character string. The name of the column containing the values for the plot. Default is third column, for subset data tables of the form .(Lon,Lat,var).}

\item{mn}{Title of the plot.}

\item{col_scheme}{Either "bwr" for blue - white - red, "wr" for white - red, or "wb" for white - blue. Specifies the color scheme of the plot.}

\item{set_white}{Forces the blue-white-red color scheme to center white at the set value if specified.}

\item{xlab, ylab}{Labeling.}

\item{save_pdf, save_dir, file_name}{Whether, where and under which name the plot should be saved.}

\item{stretch_par}{Numeric. Only used when save_pdf == TRUE. Stretches the pdf output. Default is NULL, where it is stretched to #Lons/#Lats.}

\item{legend}{Boolean. if FALSE, no legend is provided}
}
\value{
none
}
\description{
Takes a data table that contains cols Lon, Lat (or lon and lat) and a variable to plot.
}
\examples{
\dontrun{
dt = load_combined_wide()
plot_diagnostic(dt)
}

}
\author{
Claudio Heinrich
}
