% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_quantiles.R
\name{plot_hubverse_quantiles}
\alias{plot_hubverse_quantiles}
\title{Plot hubverse formatted forecasts
for all or a subset of forecasted
locations.}
\usage{
plot_hubverse_quantiles(
  forecast_data_path,
  locations = NULL,
  observed_data_path = NULL,
  start_date = NULL,
  end_date = NULL,
  location_input_format = "abbr",
  location_output_format = "abbr",
  ytrans = "log10",
  linewidth = 2,
  pointsize = 4,
  forecast_linecolor = "darkblue",
  forecast_pointcolor = "darkblue",
  obs_linecolor = "black",
  obs_pointcolor = "black",
  autotitle = TRUE
)
}
\arguments{
\item{forecast_data_path}{path to hubverse-
formatted forecast data, as a single \code{.csv}.}

\item{locations}{set of locations to plot. If NULL,
all locations are plotted. Otherwise, a vector
of location values to plot, as USPS-style
abbreviations (e.g. \verb{c("US", "AL", "AK"}),
US hubverse submission location codes (
e.g. \code{c("US", 01, 02)}), or full English
jurisdiction names
(e.g. \verb{c("United States, "Alabama", "Alaska")}.
Default \code{NULL}.}

\item{observed_data_path}{path to observed data
to plot alongside the forecast quantiles. If \code{NULL},
only the forecast quantiles will be plotted. Default \code{NULL}.}

\item{start_date}{first date to plot. If NULL, defaults
to the earliest date found between the forecast timeseries (
obtained from \code{forecast_data_path}) and the observed
data timeseries (obtained from \code{observed_data_path}, if
provided). Default NULL.}

\item{end_date}{final date to plot. If NULL, defaults
to the latest date found between the forecast timeseries (
obtained from \code{forecast_data_path}) and the observed
data timeseries (obtained from \code{observed_data_path}, if
provided). Default NULL.}

\item{location_input_format}{format of the provided location
vector. Permitted formats are \code{"abbr"} (state/territory
or nation two letter USPS abbreviation), \code{"hub"} (
legacy 2-digit FIPS code for states and territories, \code{US}
for the USA as a whole), and \code{"long_name"} (full English
jurisdiction names; not recommended). Default \code{"abbr"}.}

\item{location_output_format}{Location format for naming the
entries of the output list. Accepts the same string
keys as \code{location_input_format}.}

\item{linewidth}{\code{linewidth} parameter passed to
\code{\link[ggplot2:geom_path]{ggplot2::geom_line()}}. Default \code{2}.}

\item{pointsize}{\code{size} parameter passed to
\code{\link[ggplot2:geom_point]{ggplot2::geom_point()}}. Default \code{4}.}

\item{forecast_linecolor}{\code{color} parameter
passed to \code{\link[ggplot2:geom_path]{ggplot2::geom_line()}} for plotting forecasts.
Default \code{"darkblue"}.}

\item{forecast_pointcolor}{\code{color} parameter passed to
\code{\link[ggplot2:geom_point]{ggplot2::geom_point()}} for plotting forecasts.
Default \code{"darkblue"}.}

\item{obs_linecolor}{\code{color} parameter passed to \code{\link[ggplot2:geom_path]{ggplot2::geom_line()}}
for plotting observed data. Default \code{"black"}.}

\item{obs_pointcolor}{\code{color} parameter passed to
\code{\link[ggplot2:geom_point]{ggplot2::geom_point()}} for plotting observed data. Default \code{"black"}.}

\item{autotitle}{Generate a title for the individual
plots from the hubverse \code{reference_date} and the location
name? Boolean, default \code{TRUE}.}

\item{y_transform}{axis transform passed as the \code{transform}
argument to \code{\link[ggplot2:scale_continuous]{ggplot2::scale_y_continuous()}}. Default \code{"log10"}.}
}
\value{
a list of ggplot objects of the plots created,
one for each location
}
\description{
Plot quantiles given hubverse
formatted data for all
locations in the dataset or
a subset of them.
}
