% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_quantiles.R
\name{plot_hubverse_quantiles_loc}
\alias{plot_hubverse_quantiles_loc}
\title{Plot hubverse formatted quantile forecasts for a given
location.}
\usage{
plot_hubverse_quantiles_loc(
  location,
  forecast_data,
  truth_data,
  location_format = "abbr",
  ytrans = "log10",
  linesize = 2,
  pointsize = 4,
  forecast_linecolor = "darkblue",
  forecast_pointcolor = "darkblue",
  truth_linecolor = "black",
  truth_pointcolor = "black",
  target_name = NULL,
  autotitle = TRUE
)
}
\arguments{
\item{location}{location to plot}

\item{forecast_data}{hubverse format quantile
forecast data, as a \code{\link[tibble:tibble]{tibble}}.}

\item{truth_data}{hubverse format truth data,
as a \code{\link[tibble:tibble]{tibble}}.}

\item{location_format}{format of the provided location.
Permitted formats are \code{"abbr"} (state/territory
or nation two letter USPS abbreviation), \code{"hub"} (
legacy 2-digit FIPS code for states and territories, \code{US}
for the USA as a whole), and \code{"long_name"} (full English
jurisdiction names; not recommended). Default \code{"abbr"}.}

\item{ytrans}{axis transform passed to
\code{\link[ggplot2:scale_continuous]{ggplot2::scale_y_continuous()}}. Default \code{'log10'}.}

\item{linesize}{\code{size} parameter passed to
\code{\link[ggplot2:geom_path]{ggplot2::geom_line()}}. Default 2.}

\item{pointsize}{\code{size} parameter passed to
\code{\link[ggplot2:geom_point]{ggplot2::geom_point()}}. Default 4.}

\item{forecast_linecolor}{\code{color} parameter passed to
\code{\link[ggplot2:geom_path]{ggplot2::geom_line()}} for plotting forecasts.
Default "darkblue".}

\item{forecast_pointcolor}{\code{color} parameter passed to
\code{\link[ggplot2:geom_point]{ggplot2::geom_point()}} for plotting forecasts.
Default "darkblue".}

\item{truth_linecolor}{\code{color} parameter passed to
\code{\link[ggplot2:geom_path]{ggplot2::geom_line()}} for plotting truth data.
Default "black".}

\item{truth_pointcolor}{\code{color} parameter passed to
\code{\link[ggplot2:geom_point]{ggplot2::geom_point()}} for plotting truth data.
Default "black".}

\item{target_name}{Name of the forecast target,
for labeling the plot y-axis. Default \code{NULL},
in which case a value from the \code{target} column
in \code{forecast_data} will be used.}

\item{autotitle}{Boolean. Generate an automatic
title for the plot from the location name
and reference date? Default \code{TRUE}.}
}
\value{
the plot as a ggplot object
}
\description{
Given a tibble of data properly formatted
for a hubverse schema forecast hub submission,
plot a timeseries quantile predictions for
a given location.
}
