% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_smooth}
\alias{plot_smooth}
\title{smooth plotting function}
\usage{
plot_smooth(
  dt_plot,
  var = colnames(dt)[3],
  mn = var,
  rr = NULL,
  theta = 0.5,
  pixels = 256,
  col_scheme = "bwr",
  set_white = NULL,
  xlab = "Lon",
  ylab = "Lat",
  save_pdf = FALSE,
  save_dir = "./figures/",
  file_name = "diag_plot",
  stretch_par = NULL,
  exclude_land = FALSE,
  exclude_ocean = FALSE,
  legend = TRUE
)
}
\arguments{
\item{var}{Character string. The name of the column containing the values for the plot. Default is the name of the third column of dt, to be used for subset data tables of the form .(Lon,Lat,var).}

\item{mn, rr}{Title and range of the plot.}

\item{theta}{parameter for the Gaussian smoothing kernel}

\item{pixels}{Resolution of the plot}

\item{col_scheme}{Either "bwr" for blue - white - red, "wr" for white - red, or "wb" for white - blue. Specifies the color scheme of the plot.}

\item{set_white}{Forces the blue-white-red color scheme to center white at the set value if specified.}

\item{xlab, ylab}{Labeling.}

\item{save_pdf, save_dir, file_name}{Whether, where and under which name the plot should be saved.}

\item{stretch_par}{Numeric. Only used when save_pdf == TRUE. Stretches the pdf output. Default is NULL, where it is stretched to #Lons/#Lats.}

\item{exclude_land}{Should pixels on land be grayed out?}

\item{exclude_ocean}{Should pixels on ocean be grayed out?}

\item{legend}{Should the plot have a legend?}

\item{dt}{The data table.}
}
\value{
none
}
\description{
Takes a data table of the form .(Lon,Lat,value) or .(Lat,Lon,value) and plots values on the globe after applying a kernel smoothing
}
\examples{
\dontrun{
dt = load_combined_wide()
plot_smooth(dt)
}

}
\author{
Claudio Heinrich
}
