% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{plots_to_pdf}
\alias{plots_to_pdf}
\title{Save a list of ggplot plots as a PDF, with a
grid of \code{nrow} by \code{ncol} plots per page}
\usage{
plots_to_pdf(
  list_of_plots,
  save_path,
  nrow = 1,
  ncol = 1,
  width = 8.5,
  height = 11
)
}
\arguments{
\item{list_of_plots}{List of plots to save to PDF.}

\item{save_path}{Path to which to save the plots. Must
end in \code{.pdf}.}

\item{nrow}{Number of rows of plots per page
(passed to \code{\link[gridExtra:marrangeGrob]{gridExtra::marrangeGrob()}})
Default \code{1}.}

\item{ncol}{Number of columns of plots per page
(passed to \code{\link[gridExtra:marrangeGrob]{gridExtra::marrangeGrob()}}).
Default \code{1}.}

\item{width}{page width in device units (passed to
\code{\link[ggplot2:ggsave]{ggplot2::ggsave()}}). Default \code{8.5}.}

\item{height}{page height in device units (passed to
\code{\link[ggplot2:ggsave]{ggplot2::ggsave()}}). Default \code{11}.}
}
\value{
Nothing, saving the plots as a side effect.
}
\description{
Save a list of ggplot plots as a PDF, with a
grid of \code{nrow} by \code{ncol} plots per page
}
