% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{prism_rate_reference_populations}
\alias{prism_rate_reference_populations}
\title{Reference populations underlying PRISM rate
cutpoints.}
\format{
A tibble with 3 columns:
\describe{
\item{location}{Two-letter location abbreviation, lowercase}
\item{as_of}{Date the reference population took effect}
\item{population}{Population denominator used by PRISM}
}
}
\source{
data-raw/prism_thresholds.R
}
\usage{
prism_rate_reference_populations
}
\description{
The population denominators PRISM used to derive its
rate-scale activity level cutpoints, vintaged by
\code{as_of} in the same way as \link{prism_thresholds}.
}
\seealso{
\code{\link[=get_prism_reference_population]{get_prism_reference_population()}},
\link{prism_thresholds}
}
\keyword{datasets}
