% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_data_cdc_gov.R
\name{pull_data_cdc_gov_dataset}
\alias{pull_data_cdc_gov_dataset}
\title{Pull a dataset from \code{data.cdc.gov} with standard selection
and filtering options.}
\usage{
pull_data_cdc_gov_dataset(
  dataset,
  dataset_lookup_format = "key",
  api_key_id = Sys.getenv("DATA_CDC_GOV_API_KEY_ID"),
  api_key_secret = Sys.getenv("DATA_CDC_GOV_API_KEY_SECRET"),
  start_date = NULL,
  end_date = NULL,
  columns = NULL,
  locations = NULL,
  order_by = NULL,
  desc = FALSE,
  limit = 1e+05,
  error_on_limit = TRUE,
  rename_columns = FALSE,
  ...
)
}
\arguments{
\item{dataset}{Dataset key or id (as one of the keys or ids
in \link{data_cdc_gov_dataset_table}. Format determined by the
value of \code{dataset_lookup_format}.}

\item{dataset_lookup_format}{Format for the \code{dataset}
string. One of \code{"key"} or \code{"id"}. Default \code{"key"}.
See \code{\link[=data_cdc_gov_dataset_lookup]{data_cdc_gov_dataset_lookup()}}.}

\item{api_key_id}{Key ID of an API key to use
when querying the dataset. Not required,
but polite and reduces throttling.
You can create one at
\href{https://data.cdc.gov/profile/edit/developer_settings}{\code{data.cdc.gov/profile/edit/developer_settings} }.
Defaults to the value of the environment variable
\code{DATA_CDC_GOV_API_KEY_ID}, if any.}

\item{api_key_secret}{Associated key secret
for the API key given in \code{api_key_id}.
Defaults to the value of the environment variable
\code{DATA_CDC_GOV_API_KEY_SECRET}, if any.}

\item{start_date}{Pull only rows with dates
greater than or equal to this date. If \code{NULL},
no minimum date. Default \code{NULL}.}

\item{end_date}{Pull only rows with dates
less than or equal to this date. If \code{NULL},
no maximum date. Default \code{NULL}.}

\item{columns}{Vector of columns to retrieve, in
addition to the location and date columns for the dataset, which
are always retrieved if they exist. Default \code{NULL}.}

\item{locations}{value or values to filter on for the dataset's
location column. If \code{NULL}, do not filter on that column.
Default \code{NULL}.}

\item{order_by}{column or columns to order (sort) by.
If \code{NULL} (default) will order first by the date column
and then by the location column.}

\item{desc}{Boolean. Whether to order descending instead of
ascending. Default \code{FALSE} (order ascending).}

\item{limit}{maximum number of rows to return. Default \code{1e5}
(100000)}

\item{error_on_limit}{Boolean. Raise an error if the number
of rows returned is equal to the maximum? Default \code{TRUE}.
This ensures that one does not silently end up with a
subset of the total set of rows matching the query. If a subset
is desired, one can set \code{error_on_limit = FALSE}.}

\item{rename_columns}{Boolean. Rename the dataset-specific
date and location columns to \code{date} and \code{location}, respectively?
Default \code{FALSE}.}

\item{...}{other arguments passed to \code{\link[=data_cdc_gov_soda_query]{data_cdc_gov_soda_query()}}}
}
\value{
the pulled data, as a \code{\link[tibble:tibble]{tibble}}.
}
\description{
Pull a dataset from \code{data.cdc.gov} with standard selection
and filtering options.
}
