% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_data_cdc_gov.R
\name{pull_nhsn_hrd}
\alias{pull_nhsn_hrd}
\title{Pull NHSN Hospital Respiratory data from \code{data.cdc.gov}}
\usage{
pull_nhsn_hrd(
  release = "prelim",
  api_key_id = Sys.getenv("DATA_CDC_GOV_API_KEY_ID"),
  api_key_secret = Sys.getenv("DATA_CDC_GOV_API_KEY_SECRET"),
  start_date = NULL,
  end_date = NULL,
  columns = NULL,
  locations = NULL,
  order_by = c("jurisdiction", "weekendingdate"),
  desc = FALSE,
  limit = 1e+05,
  error_on_limit = TRUE,
  ...
)
}
\arguments{
\item{release}{HRD release to pull.
One of \code{"prelim"} (preliminary weekly data release
typically occurs on Wednesdays) or \code{"final"}
final weekly data release, typically occurs on Fridays. Default
\code{"prelim"}.}

\item{api_key_id}{Key ID of an API key to use
when querying the dataset. Not required,
but polite and reduces throttling.
You can create one at
\href{https://data.cdc.gov/profile/edit/developer_settings}{\code{data.cdc.gov/profile/edit/developer_settings} }.
Defaults to the value of the environment variable
\code{DATA_CDC_GOV_API_KEY_ID}, if any.}

\item{api_key_secret}{Associated key secret
for the API key given in \code{api_key_id}.
Defaults to the value of the environment variable
\code{DATA_CDC_GOV_API_KEY_SECRET}, if any.}

\item{start_date}{Pull only rows with dates
greater than or equal to this date. If \code{NULL},
no minimum date. Default \code{NULL}.}

\item{end_date}{Pull only rows with dates
less than or equal to this date. If \code{NULL},
no maximum date. Default \code{NULL}.}

\item{columns}{Vector of columns to retrieve, in
addition to \code{weekendingdate} and \code{jurisdiction}, which are always
retrieved. If \code{NULL}, retrieve all columns.
Default \code{NULL}.}

\item{order_by}{column or columns to order (sort) by.
Default \code{c("jurisdiction", "weekendingdate")}
(sort first by jurisdiction, then by date).}

\item{desc}{Boolean. Whether to order descending instead of
ascending. Default \code{FALSE} (order ascending).}

\item{limit}{maximum number of rows to return. Default \code{1e5}
(100000)}

\item{error_on_limit}{Boolean. Raise an error if the number
of rows returned is equal to the maximum? Default \code{TRUE}.
This ensures that one does not silently end up with a
subset of the total set of rows matching the query. If a subset
is desired, one can set \code{error_on_limit = FALSE}.}

\item{...}{other arguments passed to \code{\link[=nhsn_soda_query]{nhsn_soda_query()}}}

\item{jurisdictions}{value or values to filter on for the \code{jurisdiction}
column of the NHSN dataset. If \code{NULL}, do not filter on that column.
Default \code{NULL}.}
}
\value{
the pulled data, as a \code{\link[tibble:tibble]{tibble}}.
}
\description{
Defaults to the \href{https://data.cdc.gov}{\code{data.cdc.gov}}
public API endpoint for Wednesday preliminary
releases.
}
