% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hubverse_to_scorable.R
\name{quantile_table_to_scorable}
\alias{quantile_table_to_scorable}
\title{Create a \code{scoringutils}-ready quantile table from a hubverse table}
\usage{
quantile_table_to_scorable(
  hubverse_quantile_table,
  observation_table,
  obs_value_column = "value",
  obs_date_column = "date",
  id_cols = c("location", "target"),
  quantile_tol = 10
)
}
\arguments{
\item{hubverse_quantile_table}{quantile forecasts,
as a hubverse-format \code{\link[tibble:tibble]{tibble}}, e.g.
as produced by \code{\link[=get_hubverse_quantile_table]{get_hubverse_quantile_table()}}, with columns including
\code{location}, \code{target_end_date}, \code{output_type}, \code{output_type_id},
and \code{value}.}

\item{observation_table}{observations, as a
\code{\link[tibble:tibble]{tibble}}.}

\item{obs_value_column}{Name of the column containing
observed values in the \code{observed} table, as a string.
Default \code{"value"}}

\item{obs_date_column}{Name of the column containing
date values in the \code{observed} table, as a string.
Default \code{"date"}.}

\item{id_cols}{Additional id columns for joining the
\verb{observation table} to the \code{hubverse_quantile_table}.
Passed to \code{\link[=hubverse_table_with_obs]{hubverse_table_with_obs()}}. Default
\code{c("location", "target")}.}

\item{quantile_tol}{Round quantile level values to this many
decimal places, to avoid problems with floating point number
equality comparisons. Passed as the \code{digits} argument to
\code{\link[base:round]{base::round()}}. Default 10.}
}
\value{
A \code{\link[data.table:data.table]{data.table}} for scoring,
as the output of \code{\link[scoringutils:as_forecast_quantile]{scoringutils::as_forecast_quantile()}}.
}
\description{
Expects quantile forecast output in hubverse format
(e.g. as created by \code{\link[=get_hubverse_quantile_table]{get_hubverse_quantile_table()}})
and an observed data table with location, date, and value columns.
The column names in the observed data table can be configured;
defaults are \code{"location"}, \code{"date"}, and
\code{"value"}, respectively.
}
