% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_bottom_up.R
\name{rank_sampled_trajectories}
\alias{rank_sampled_trajectories}
\title{Rank sampled trajectories}
\usage{
rank_sampled_trajectories(
  base_forecasts,
  location_col,
  draw_col,
  rank_quantity_col
)
}
\arguments{
\item{base_forecasts}{A data frame containing the base forecasts.}

\item{location_col}{The name of the column representing the location.}

\item{draw_col}{The name of the column representing the draw.}

\item{rank_quantity_col}{The name of the column representing the quantity
to rank trajectories by.}
}
\value{
A data frame with the ranked base forecasts.
}
\description{
This function ranks the base forecasts based on a specified quantity,
recorded in the \code{rank_quantity_col} column.
grouped by location and draw.
}
