% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_bottom_up.R
\name{sample_aggregated_trajectories}
\alias{sample_aggregated_trajectories}
\title{Generate sample trajectories based on ranked base forecasts}
\usage{
sample_aggregated_trajectories(
  base_forecasts,
  samples,
  ranked_base_forecasts,
  number_of_sampled_trajs,
  location_col,
  draw_col,
  date_col,
  value_to_aggregate_col
)
}
\arguments{
\item{base_forecasts}{A data frame containing the base forecasts.}

\item{samples}{A tibble containing the samples from a copula by
location.}

\item{ranked_base_forecasts}{A tibble containing the ranked base forecasts}

\item{number_of_sampled_trajs}{A tibble containing the number of
samples per forecast trajectory for each location.}

\item{location_col}{The name of the column representing the location}

\item{draw_col}{The name of the column representing the draw}

\item{date_col}{The name of the column representing the date}

\item{value_to_aggregate_col}{The name of the column representing
the value to aggregate}
}
\value{
A data frame containing the sampled forecasts
}
\description{
This function takes in a set of copula samples, ranked base forecasts,
and column names to generate sampled trajectories for each location and then
sum aggregate by date over locations.
}
