% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbc_backend_arviz.R
\name{sbc_backend_arviz}
\alias{sbc_backend_arviz}
\title{SBC Backend for Arviz}
\usage{
sbc_backend_arviz(...)
}
\arguments{
\item{...}{Arguments to be passed to the backend. The argument 'data' should
not be provided as it needs to be set by the SBC package. The argument 'path'
(to the Arviz datafile) must be provided.}
}
\value{
A list with the provided arguments, structured as an SBC backend for
Arviz.
}
\description{
This function creates an SBC backend for Arviz. This backend is used to hold
the arguments that will be passed to the Arviz backend, mainly the path to
the Arviz datafile. The other purpose is to extend the S3 generics from the
SBC package.
}
\examples{
\dontrun{
sbc_backend_arviz(path = "path/to/arviz_datafile")
}

}
