% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms.R
\name{scaled_logit}
\alias{scaled_logit}
\title{Scaled logit transform}
\usage{
scaled_logit(x, a, b)
}
\arguments{
\item{x}{Value to transform}

\item{a}{Left endpoint of the domain}

\item{b}{Right endpoint of the domain}
}
\value{
Transformed value
}
\description{
Compute \deqn{f(x) = \log(x - a) - \log(b - x)}
for \eqn{b > a}. This is an increasing function that
maps the interval \eqn{(a, b)} to \eqn{(-\infty, \infty)}.
}
\examples{

# with a = 0, b = 1, equivalent to `qlogis`
scaled_logit(0.7, 0, 1)
qlogis(0.7)

# but can work with other domains
scaled_logit(3, 1, 7)

}
