% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms.R
\name{scaled_shifted_log}
\alias{scaled_shifted_log}
\title{Scaled and shifted logarithmic transform}
\usage{
scaled_shifted_log(x, m, b)
}
\arguments{
\item{x}{Value to transform}

\item{m}{Scale parameter. Most relevant for its
sign. Typical choices are \code{-1} or \code{1}.}

\item{b}{finite endpoint of the domain (whether
it is the left or right endpoint is determined
by the sign of \code{m}).}
}
\value{
Transformed value
}
\description{
Compute \deqn{f(x) = m \log(m(x - b))}
}
\details{
This is an increasing function.
When \eqn{m > 0}, it maps \eqn{(b, \infty)} to
\eqn{(-\infty, \infty)}. When \eqn{m < 0}, it maps
\eqn{(-\infty, b)} to \eqn{-\infty, \infty)}.
}
\examples{

# with m = 1, b = 0, equivalent to `log`

scaled_shifted_log(5, 1, 0)
log(5)

# but can handle other domains

scaled_shifted_log(-exp(3), -1, 5)

}
