% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{squarify}
\alias{squarify}
\title{Take a data.table on an incomplete grid and "squarify" so that the image function can work on it}
\usage{
squarify(dt, var = names(dt)[3])
}
\arguments{
\item{dt}{A data.table with Lon and Lat plus another variable}

\item{var}{The name of the variable to be squarified}
}
\description{
Take a data.table on an incomplete grid and "squarify" so that the image function can work on it
}
