% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_scores_with_baseline.R
\name{summarise_scores_with_baseline}
\alias{summarise_scores_with_baseline}
\alias{summarize_scores_with_baseline}
\title{Summarize a scoring table with scores relative to a baseline model.}
\usage{
summarise_scores_with_baseline(
  scores,
  baseline,
  compare = "model",
  metric_to_compare = intersect(c("wis", "crps", "brier_score"), names(scores)),
  by = NULL,
  ...
)

summarize_scores_with_baseline(
  scores,
  baseline,
  compare = "model",
  metric_to_compare = intersect(c("wis", "crps", "brier_score"), names(scores)),
  by = NULL,
  ...
)
}
\arguments{
\item{scores}{A scoring object containing the scoring table with
quantile scores.}

\item{baseline}{Name of the baseline, as a string. Should be
an element of the column defined by the \code{compare} argument.}

\item{compare}{Column name of the column that uniquely
identifies sets of forecasts (typically, models) to compare, as
a string. Passed as the \code{compare} argument to
\code{\link[scoringutils:get_pairwise_comparisons]{scoringutils::get_pairwise_comparisons()}}. Default \code{"model"},
matching the default of \code{\link[scoringutils:get_pairwise_comparisons]{scoringutils::get_pairwise_comparisons()}}.}

\item{metric_to_compare}{Name of the (absolute) metric for which to
compute relative values. Passed as the
\code{metric} argument to \code{\link[scoringutils:get_pairwise_comparisons]{scoringutils::get_pairwise_comparisons()}}.
By default, chooses \code{"wis"}, \code{"crps"}, or \code{"brier_score"}
if that metric is available in the input scores. This matches the
default in \code{\link[scoringutils:get_pairwise_comparisons]{scoringutils::get_pairwise_comparisons()}}.}

\item{by}{Columns besides the column in \code{compare} to group by
when scoring. \code{by} is passed directly as the \code{by} argument to
\code{\link[scoringutils:get_pairwise_comparisons]{scoringutils::get_pairwise_comparisons()}}.
\code{c(compare, by)} is passed as the \code{by} argument to
\code{\link[scoringutils:summarise_scores]{scoringutils::summarise_scores()}}.
Default \code{NULL}.}

\item{...}{additional arguments passed to
\code{\link[scoringutils:summarise_scores]{scoringutils::summarise_scores()}}.}
}
\value{
A data frame with summarised scores for each model, including
relative scores where possible.
}
\description{
Light wrapper of \code{\link[scoringutils:summarise_scores]{scoringutils::summarise_scores()}} and
\code{\link[scoringutils:get_pairwise_comparisons]{scoringutils::get_pairwise_comparisons()}} that automatically
adds relative metrics with a baseline.
}
\examples{
quantile_summary <- scoringutils::example_quantile |>
  scoringutils::score() |>
  summarise_scores_with_baseline(
  baseline = "EuroCOVIDhub-baseline")
)
print(quantile_summary)

sample_summary <- scoringutils::example_sample_discrete |>
  scoringutils::score() |>
  summarise_scores_with_baseline(
  baseline = "EuroCOVIDhub-baseline",
  by = "location")
print(sample_summary)

}
