% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sym_limits_log}
\alias{sym_limits_log}
\title{Get limits spanning a set of values
that are symmetric about a central value
in logarthmic space}
\usage{
sym_limits_log(values, center = 1)
}
\arguments{
\item{values}{Vector of values.}

\item{center}{Center value for the span.
Default 1.}
}
\value{
a length-two vector whose entries
are symmetric about \code{center} in logarithmic
space and span the range of \code{values}.
}
\description{
This is useful for making the axis range of a
plot symmetric about a key central value when
the axis scale is logarithmic.
}
\details{
Given a center point \code{center} and a set of values
\code{values}, returns \code{c(x / center, x * center)} where
\code{x} is the entry of \code{values} that is farthest
from \code{center} in logarithmic space / has the most
extreme ratio with center
(i.e. \code{x <- max(c(values / center, center / values))}).
}
\examples{
sym_limits_log(c(1, 2.2, 5))

sym_limits_log(c(3, 4, 5), center = 2)

library(ggplot2)
library(tibble)
data <- tibble(x = 1:5, y = 10^rnorm(5))
plot <- ggplot(
    data = data,
    mapping = aes(x = x, y = y)) +
    geom_point() +
    geom_hline(yintercept = 1) +
    scale_y_continuous(transform = "log10") +
    coord_cartesian(ylim = sym_limits_log(data$y))
plot

}
\seealso{
\code{\link[=sym_limits]{sym_limits()}}
}
