% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_hubverse.R
\name{target_end_dates_from_horizons}
\alias{target_end_dates_from_horizons}
\title{Get a correctly-formatted set of hubverse
\code{target_end_date} values for the given reference
date and set of horizons.}
\usage{
target_end_dates_from_horizons(
  reference_date,
  horizons,
  horizon_timescale = c("days", "weeks")
)
}
\arguments{
\item{reference_date}{hub forecasting reference date}

\item{horizons}{Vector of forecast horizons to include,
in weeks ahead of the \code{reference_date}.}

\item{horizon_timescale}{Either "days" or "weeks"}
}
\value{
a vector of target end dates
}
\description{
Get a correctly-formatted set of hubverse
\code{target_end_date} values for the given reference
date and set of horizons.
}
