% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode_locations.R
\name{to_location_table_column}
\alias{to_location_table_column}
\title{Map location format strings to corresponding
location lookup table column names.}
\usage{
to_location_table_column(location_format)
}
\arguments{
\item{location_format}{the format string to convert.}
}
\value{
the corresponding column name in
\link{us_location_table}.
}
\description{
Convert a location format string to the name of the
corresponding column in \link{us_location_table},
raising an error if an unknown format is provided.
}
