% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode_locations.R
\name{to_us_location_table_column}
\alias{to_us_location_table_column}
\alias{to_location_table_column}
\title{Map location format strings to corresponding
us location lookup table column names.}
\usage{
to_us_location_table_column(location_format)

to_location_table_column(location_format)
}
\arguments{
\item{location_format}{the format string to convert.
Valid values are:
\itemize{
\item \code{"abbr"} or \code{"short_name"}: USPS 2-letter abbreviation
\item \code{"code"} or \code{"hub"}: Legacy 2-digit FIPS code for states and
territories, \code{US} for the US as a whole, the schema typically used in
forecast hubs.
\item \code{"long_name"} or \code{"name"}: Full location name, as a string.
}}
}
\value{
the corresponding column name in
\link{us_location_table}.
}
\description{
Convert a location format string to the name of the
corresponding column in \link{us_location_table},
raising an error if an unknown format is provided.
}
