% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode_locations.R
\name{us_location_lookup}
\alias{us_location_lookup}
\alias{location_lookup}
\title{Look up rows of the USA location table
corresponding to the entries of a given
vector.}
\usage{
us_location_lookup(
  location,
  location_input_format,
  location_output_format = NULL
)

location_lookup(location, location_input_format, location_output_format = NULL)
}
\arguments{
\item{location}{vector of location values}

\item{location_input_format}{format in which the location
vector is coded. See \code{\link[=to_us_location_table_column]{to_us_location_table_column()}}
for permitted formats.}

\item{location_output_format}{Vector specifying column(s) from
the output table to return. If not provided (default),
return all columns. See \code{\link[=to_us_location_table_column]{to_us_location_table_column()}}
for permitted strings to specify columns.}
}
\value{
A \code{\link[tibble:tibble]{tibble}} with the
corresponding rows of the \link{us_location_table}
matching the location vector (with repeats possible) or the
values of those rows for given column(s), as specified in
\code{location_output_format}.
}
\description{
Get the rows of
\link{us_location_table}
corresponding to a given location vector
and format, with repeats possible
}
\examples{

us_location_lookup(c("01", "05", "US", "05"), "code")

us_location_lookup(c("01", "05", "US", "05"), "code", "name")

us_location_lookup(c("01", "05", "US", "05"), "code", c("abbr", "name"))

us_location_lookup(c("Alaska", "Hawaii"), "name", c("code", "abbr"))

}
\seealso{
\code{\link[=us_location_recode]{us_location_recode()}}
}
