% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode_locations.R
\name{us_location_recode}
\alias{us_location_recode}
\title{Recode a vector of US locations
from one format to another}
\usage{
us_location_recode(location, location_input_format, location_output_format)
}
\arguments{
\item{location}{vector of location values}

\item{location_input_format}{format in which the location
vector is coded. See \code{\link[=to_us_location_table_column]{to_us_location_table_column()}}
for permitted formats. Must be a single value.}

\item{location_output_format}{Code the output vector in this
format. See \code{\link[=to_us_location_table_column]{to_us_location_table_column()}}
for permitted formats. Must be a single value.}
}
\value{
The recoded vector.
}
\description{
Uses \code{\link[=us_location_lookup]{us_location_lookup()}} to perform recoding.
}
\examples{

us_location_recode(c("01", "05", "US", "05"), "code", "name")

us_location_recode(c("AK", "US", "HI", "AK"), "abbr", "code")

us_location_recode("AK", "abbr", "name")

}
