% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{us_location_table}
\alias{us_location_table}
\title{Table of US location names and codes}
\format{
A tibble with 58 rows and 4 columns:
\describe{
\item{code}{2-character location code}
\item{abbr}{2-character location abbreviation}
\item{hrd}{location abbreviation abbreviation as in the \code{jurisdiction} column of the NHSN Hospital Respiratory Data (HRD) dataset. Same as \code{abbr} for subnational jurisdictions, but \code{USA} for the United States as a whole.}
\item{name}{Full name of the location}
}
}
\source{
\url{https://www2.census.gov/geo/docs/reference/state.txt}

data-raw/us_location_table.R
}
\usage{
us_location_table
}
\description{
A table of US location names and codes, assembled from a
US census reference table with the addition of "US" for the
United States as a whole.
}
\keyword{datasets}
