% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_bottom_up.R
\name{validate_base_forecasts}
\alias{validate_base_forecasts}
\title{Validate the base forecasts}
\usage{
validate_base_forecasts(
  base_forecasts,
  cp,
  draw_col = ".draw",
  date_col = "date",
  value_to_aggregate_col = "hosp",
  rank_quantity_col = "hosp",
  location_col = "location"
)
}
\arguments{
\item{base_forecasts}{The base forecasts tibble}

\item{cp}{Base forecast sampling copula. This expects
a copula object from the \code{copula} package, for example
\code{copula::normalCopula(0.5, dim = 53)}.}

\item{draw_col}{String name of column for draws.}

\item{date_col}{String name of column for dates.}

\item{value_to_aggregate_col}{String name of column to make bottom-up
aggregation on.}

\item{rank_quantity_col}{String name of column to rank draw/location
pair on.}

\item{location_col}{String name of the location column.}
}
\value{
Nothing.
}
\description{
Validate the base forecasts tibble. This is aimed at, first, checking
that columns exist as expected and, second, that it is possible to sample
trajectories from the base forecast tibble by matching a draw index for each
location along an ascending sequence of dates.
}
