% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hubverse_to_scorable.R
\name{with_hubverse_oracle_output}
\alias{with_hubverse_oracle_output}
\title{Join hubverse oracle output to a hubverse forecast table for scoring.}
\usage{
with_hubverse_oracle_output(hubverse_table, oracle_output_table)
}
\arguments{
\item{hubverse_table}{Hubverse format table of forecasts}

\item{oracle_output_table}{Oracle output table to join to the forecast table.}
}
\value{
The joined table, as a \code{\link[tibble:tibble]{tibble}}.
}
\description{
Join hubverse oracle output to a hubverse forecast table for scoring.
}
\examples{

with_hubverse_oracle_output(hubExamples::forecast_outputs,
  hubExamples::forecast_oracle_output)

}
