% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_tabular_file}
\alias{read_tabular_file}
\alias{write_tabular_file}
\title{Read from or write to tabular files, with
file format inferred from the file extension.}
\usage{
read_tabular_file(path_to_file, ...)

write_tabular_file(table, path_to_file, ...)
}
\arguments{
\item{path_to_file}{Path to the file to read/write.}

\item{...}{Additional keyword arguments passed to the
file reader/writer function, which will be one of
\code{\link[readr:read_delim]{readr::read_csv()}} / \code{\link[readr:write_delim]{readr::write_csv()}},
\code{\link[readr:read_delim]{readr::read_tsv()}} / \code{\link[readr:write_delim]{readr::write_tsv()}}, and
\code{\link[arrow:read_parquet]{arrow::read_parquet()}} / \code{\link[arrow:write_parquet]{arrow::write_parquet()}},
depending on the \code{file_format}.}

\item{table}{Table to write (\code{write_tabular_file} only).}
}
\value{
For \code{read_tabular_file}, the result of
reading in the file, as a
\code{\link[tibble:tibble]{tibble}}. For \code{write_tabular_file},
nothing, saving the tabular to disk as a side effect.
}
\description{
Read from or write to tabular files, with
file format inferred from the file extension.
}
