% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorize_prism.R
\name{categorize_prism}
\alias{categorize_prism}
\title{Categorize a numeric vector into PRISM
activity level bins.}
\usage{
categorize_prism(
  value,
  location,
  disease,
  as_of = lubridate::today(),
  prism_bin_names = NULL,
  signal = lifecycle::deprecated()
)
}
\arguments{
\item{value}{numeric vector to categorize}

\item{location}{vector of location of length equal
to \code{value} or a single location for all \code{value}.}

\item{disease}{vector of disease of length equal to
\code{value} or a single disease for all \code{value}.}

\item{as_of}{Retrieve cutpoints that were in place as of
this date. Defaults to today (current cuptoints).}

\item{prism_bin_names}{Bin names for the PRISM bins,
in order from lowest to highest. Must be a vector of
length 5. \code{list(prism_bin_names)} will be passed as
the \code{label_sets} argument to \code{\link[=categorize_vector]{categorize_vector()}}.
If \code{NULL} (the default), derived from the cutpoint
names by dropping the upper bound and converting to
title case, giving \code{"Very Low"}, \code{"Low"},
\code{"Moderate"}, \code{"High"}, and \code{"Very High"}.}

\item{signal}{surveillance signal for which to
return the cutpoints. Options are \code{"NSSP"} (proportions
of emergency department visits) and \code{"NHSN"} (weekly
hospital admissions per 100k population).
If not specified, default to \code{"NSSP"} with a
deprecation warning.}
}
\value{
A factor vector of category labels, equal in
length to the input vector \code{value}.
}
\description{
Uses \code{\link[=categorize_vector]{categorize_vector()}} and \code{\link[=get_prism_cutpoints]{get_prism_cutpoints()}}.
}
