% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorize_prism.R
\name{categorize_prism}
\alias{categorize_prism}
\title{Categorize a vector of value into PRISM
activity level bins.}
\usage{
categorize_prism(
  value,
  location,
  disease,
  as_of = lubridate::today(),
  prism_bin_names = default_prism_bin_names,
  signal = lifecycle::deprecated()
)
}
\arguments{
\item{value}{value to categorize}

\item{location}{vector of location of length equal
to \code{value} or a single location for all \code{value}.}

\item{disease}{vector of disease of length equal to
\code{value} or a single disease for all \code{value}.}

\item{as_of}{date for which to get the PRISM
cutpoints. Defaults to today.}

\item{prism_bin_names}{Bin names for the PRISM bins.
in order from lowest to highest. Must be a vector of
length 5. \code{list(prism_bin_names)} will be passed as
the \code{label_sets} argument to \code{\link[=categorize_vector]{categorize_vector()}}.
Defaults to the standard PRISM bin names in title
case: \code{default_prism_bin_names}.}

\item{signal}{surveillance signal for which to
return the cutpoints. One of \code{"NSSP"} (proportions
of emergency department visits) or \code{"NHSN"} (weekly
hospital admissions per 100k population). If not
given, defaults to \code{"NSSP"} with a deprecation
warning (a future version will require it).}
}
\value{
A factor vector of category labels, equal in
length to the input vector \code{value}.
}
\description{
Uses \code{\link[=categorize_vector]{categorize_vector()}} and \code{\link[=get_prism_cutpoints]{get_prism_cutpoints()}}.
}
