% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorize_prism.R
\name{categorize_prism}
\alias{categorize_prism}
\title{Categorize a vector of values into PRISM
activity level bins.}
\usage{
categorize_prism(
  values,
  locations,
  diseases,
  prism_bin_names = c("Very Low", "Low", "Moderate", "High", "Very High")
)
}
\arguments{
\item{values}{values to categorize}

\item{locations}{vector of locations of length equal to
\code{values} or a single location for all \code{values}.}

\item{diseases}{vector of diseases of length equal to
\code{values} or a single disease for all \code{values}.}

\item{prism_bin_names}{Bin names for the PRISM bins.
in order from lowest to highest. Must be a vector of
length 5. \code{list(prism_bin_names)} will be passed as the
\code{label_sets} argument to \code{\link[=categorize_vector]{categorize_vector()}}.
Defaults to the standard PRISM bin names in title case:
\code{c("Very Low", "Low", "Moderate", "High", "Very High")}.}
}
\value{
A factor vector equal in length to \code{values} of
the categories, as the output of \code{\link[=categorize_vector]{categorize_vector()}}.
}
\description{
Categorize a vector of values into PRISM
activity level bins.
}
