% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorize_prism.R
\name{categorize_prism}
\alias{categorize_prism}
\title{Categorize a vector of values into PRISM
activity level bins.}
\usage{
categorize_prism(
  values,
  locations,
  diseases,
  prism_bin_names = default_prism_bin_names
)
}
\arguments{
\item{values}{values to categorize}

\item{locations}{vector of locations of length equal to
\code{values} or a single location for all \code{values}.}

\item{diseases}{vector of diseases of length equal to
\code{values} or a single disease for all \code{values}.}

\item{prism_bin_names}{Bin names for the PRISM bins.
in order from lowest to highest. Must be a vector of
length 5. \code{list(prism_bin_names)} will be passed as the
\code{label_sets} argument to \code{\link[=categorize_vector]{categorize_vector()}}.
Defaults to the standard PRISM bin names in title case:
\code{default_prism_bin_names}.}
}
\value{
A factor vector of category labels, equal in length
to the input vector \code{values}.
}
\description{
Uses \code{\link[=categorize_vector]{categorize_vector()}} and \code{\link[=get_prism_cutpoints]{get_prism_cutpoints()}}.
}
