% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{cdc_hub_additional_colnames}
\alias{cdc_hub_additional_colnames}
\title{Names of additional standard columns for
the CDC FluSight and COVID-19 Forecast Hubs
beyond those in \link{hubverse_std_colnames}.}
\format{
An object of class \code{character} of length 5.
}
\usage{
cdc_hub_additional_colnames
}
\description{
Names of additional standard columns for
the CDC FluSight and COVID-19 Forecast Hubs
beyond those in \link{hubverse_std_colnames}.
}
\keyword{datasets}
