% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{cdc_hub_std_colnames}
\alias{cdc_hub_std_colnames}
\alias{covidhub_std_colnames}
\alias{flusight_std_colnames}
\title{Names of standard columns for the
the CDC FluSight and COVID-19 Forecast Hubs}
\format{
An object of class \code{character} of length 9.

An object of class \code{character} of length 9.

An object of class \code{character} of length 9.
}
\usage{
cdc_hub_std_colnames

covidhub_std_colnames

flusight_std_colnames
}
\description{
Names of standard columns for the
the CDC FluSight and COVID-19 Forecast Hubs

Alias for \link{cdc_hub_std_colnames}.

Alias for \link{cdc_hub_std_colnames}.
}
\keyword{datasets}
