% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daily_to_epiweekly.R
\name{daily_to_epiweekly}
\alias{daily_to_epiweekly}
\title{Aggregate a daily timeseries to epiweekly}
\usage{
daily_to_epiweekly(
  tidy_daily_trajectories,
  value_col = "value",
  date_col = "date",
  id_cols = ".draw",
  weekly_value_name = "weekly_value",
  strict = TRUE
)
}
\arguments{
\item{tidy_daily_trajectories}{tibble of trajectories,
with dates, trajectory ids, and observeable values to
aggregate}

\item{value_col}{name of the column containing trajectory
values. Default \code{"value"}.}

\item{date_col}{name of the column containing dates.
Default \code{"date"}.}

\item{id_cols}{name(s) of the column(s)
that uniquely identify a single timeseries
(e.g. a single location timeseries or a single
posterior trajectory). Default \code{".draw"} (as the output of
\code{\link[tidybayes:spread_draws]{tidybayes::spread_draws()}}).}

\item{weekly_value_name}{name to use for the output coluqmn
containing weekly trajectory values. Default \code{"weekly_value"}.}

\item{strict}{Boolean. If \code{TRUE} only aggregate an epiweek
if all seven days are represented in the trajectory.
If \code{FALSE}, allow partial weeks. Default \code{TRUE}.}
}
\value{
tibble of aggregated epiweekly trajectories.
}
\description{
Given a set of daily timeseries / time trajectories in
tidy long-form format (i.e. with one row per day
per trajectory, with one or more columns that uniquely
identify individual trajectories (e.g. a \code{trajectory_id} column),
aggregate it to the scale of epiweeks.
}
\details{
This requires a column containing a set of timepoints
coercible to dates.
}
