% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_data_cdc_gov.R
\name{data_cdc_gov_dataset_id}
\alias{data_cdc_gov_dataset_id}
\title{Retrieve the dataset IDs for supported \code{data.cdc.gov} datasets}
\usage{
data_cdc_gov_dataset_id(dataset_key)
}
\arguments{
\item{dataset_key}{Internal forecasttools name for the dataset.
See \link{data_cdc_gov_dataset_table}.}
}
\value{
The dataset ids, as a vector string.
}
\description{
Retrieve the dataset IDs for supported \code{data.cdc.gov} datasets
}
\examples{
data_cdc_gov_dataset_id("nhsn_hrd_prelim")

data_cdc_gov_dataset_id("nssp_prop_ed_visits")

data_cdc_gov_dataset_id(c("nssp_prop_ed_visits", "nhsn_hrd_final"))

}
