% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_data_cdc_gov.R
\name{data_cdc_gov_endpoint}
\alias{data_cdc_gov_endpoint}
\title{Construct a data.cdc.gov API endpoint from a dataset ID}
\usage{
data_cdc_gov_endpoint(dataset_id)
}
\arguments{
\item{dataset_id}{Dataset ID, as a string.}
}
\value{
The full URL for the (JSON) Socrata Open Data API endpoint.
}
\description{
Construct a data.cdc.gov API endpoint from a dataset ID
}
\examples{

data_cdc_gov_endpoint(data_cdc_gov_ids$nhsn_hrd_prelim)

data_cdc_gov_endpoint(data_cdc_gov_ids$nssp_prop_ed_visits)
}
