% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{.warn_no_api_creds}
\alias{.warn_no_api_creds}
\title{Helper functions for interacting with APIs
Warn the user that they are attempting an API request
without credentials}
\usage{
.warn_no_api_creds(api_key_creation_url = NULL)
}
\arguments{
\item{api_key_creation_url}{Optional URL to which to direct
the user to create an API key ID / secret pair.}
}
\value{
Nothing, invisibly, warning the user as a side effect.
}
\description{
Helper functions for interacting with APIs
Warn the user that they are attempting an API request
without credentials
}
\keyword{internal}
