% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_path_point.R
\name{draw_key_path_point}
\alias{draw_key_path_point}
\title{Legend key glyph for \link{GeomPathPoint} and subclasses of it.}
\usage{
draw_key_path_point(data, params, size)
}
\description{
Legend key glyph for \link{GeomPathPoint} and subclasses of it.
}
