% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiweek.R
\name{epiweek_fn}
\alias{epiweek_fn}
\title{Get a date-to-epiweek function from an
epiweek standard.}
\usage{
epiweek_fn(epiweek_standard)
}
\arguments{
\item{epiweek_standard}{One of \code{"MMWR"} or \code{"USA"}
(USA MMWR epiweek, starts on Sunday) and \code{"ISO"} (ISO
week, starts on Monday). Not case-sensitive.
Must be a single value.}
}
\value{
A function, either \code{\link[lubridate:week]{lubridate::epiweek()}} or
\code{\link[lubridate:week]{lubridate::isoweek()}}, depending on the value of
\code{epiweek_standard}
}
\description{
Get the function that converts dates to epiweek numbers
for a given epiweek standard.
}
