% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecasttools-package.R
\docType{package}
\name{forecasttools-package}
\alias{forecasttools}
\alias{forecasttools-package}
\title{forecasttools: R toolbox for short-term infectious disease forecasting}
\description{
A collection of tools for short-term forecasting within CFA Predict.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://cdcgov.github.io/forecasttools}
}

}
\author{
\strong{Maintainer}: Dylan Morris \email{dzl1@cdc.gov}

Authors:
\itemize{
  \item Samuel Brand \email{usi1@cdc.gov}
}

Other contributors:
\itemize{
  \item Kaitlyn Johnson \email{uox1@cdc.gov} [contributor]
  \item Damon Bayer \email{xum8@cdc.gov} [contributor]
}

}
\keyword{internal}
