% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecasttools-package.R
\docType{package}
\name{forecasttools-package}
\alias{forecasttools}
\alias{forecasttools-package}
\title{forecasttools: R toolbox for short-term infectious disease forecasting.}
\description{
A collection of tools for short-term forecasting within CFA Predict.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://potential-engine-5mz851o.pages.github.io/}
}

}
\author{
\strong{Maintainer}: Dylan Morris \email{dylan@dylanhmorris.com}

Authors:
\itemize{
  \item Samuel Brand \email{usi1@cdc.gov}
}

Other contributors:
\itemize{
  \item Kaitlyn Johnson \email{uox1@cdc.gov} [contributor]
}

}
\keyword{internal}
