% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hub_utils.R
\name{gather_hub_target_data}
\alias{gather_hub_target_data}
\title{Gather target truth data from a forecast hub.}
\usage{
gather_hub_target_data(
  hub_path,
  local_datapath = fs::path("target-data", "target-hospital-admissions.csv")
)
}
\arguments{
\item{hub_path}{Local path to hub.}

\item{local_datapath}{The local path to the truth data file for
relative to forecast hub directory.
Defaults to expected path in the FluSight forecast hub.}
}
\value{
Table of target truth data
}
\description{
Gather target truth data from a forecast hub.
}
