% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_path_point.R
\docType{data}
\name{GeomPathPoint}
\alias{GeomPathPoint}
\title{Geometric object showing a path of connected points.}
\format{
An object of class \code{GeomPathPoint} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 7.
}
\usage{
GeomPathPoint
}
\description{
Geometric object showing a path of connected points.
}
\keyword{datasets}
