% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_path_point.R
\docType{data}
\name{GeomPathPoint}
\alias{GeomPathPoint}
\alias{GeomLinePoint}
\title{Geometric object showing a path of connected points.}
\format{
An object of class \code{GeomPathPoint} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 7.

An object of class \code{GeomLinePoint} (inherits from \code{GeomPathPoint}, \code{Geom}, \code{ggproto}, \code{gg}) of length 4.
}
\usage{
GeomPathPoint

GeomLinePoint
}
\description{
Geometric object showing a path of connected points.

Geometric object showing a path of connected points sorted by
orderded in the order of the x-axis variable.
}
\keyword{datasets}
