% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_hubverse.R
\name{get_hubverse_quantile_table}
\alias{get_hubverse_quantile_table}
\alias{get_epiweekly_hubverse_table}
\alias{get_flusight_hub_table}
\alias{get_covid_hub_table}
\title{Format quantile forecasts for a hubverse submission}
\usage{
get_hubverse_quantile_table(
  quantile_forecasts,
  reference_date,
  target_name,
  quantile_value_col = "quantile_value",
  quantile_level_col = "quantile_level",
  location_col = "location",
  timepoint_col = "target_end_date",
  target_end_dates = NULL,
  horizons = NULL,
  horizon_timescale = c("days", "weeks"),
  excluded_locations = NULL,
  quantile_tol = 4
)

get_epiweekly_hubverse_table(
  quantile_forecasts,
  reference_date,
  target_name,
  quantile_value_col = "quantile_value",
  quantile_level_col = "quantile_level",
  location_col = "location",
  timepoint_col = "target_end_date",
  target_end_dates = NULL,
  horizons = NULL,
  reference_dow = NULL,
  week_start = NULL,
  excluded_locations = NULL,
  quantile_tol = 4
)

get_flusight_hub_table(
  quantile_forecasts,
  reference_date,
  target_name,
  quantile_value_col = "quantile_value",
  quantile_level_col = "quantile_level",
  location_col = "location",
  timepoint_col = "target_end_date",
  quantile_tol = 4
)

get_covid_hub_table(
  quantile_forecasts,
  reference_date,
  target_name,
  quantile_value_col = "quantile_value",
  quantile_level_col = "quantile_level",
  location_col = "location",
  timepoint_col = "target_end_date",
  quantile_tol = 4
)
}
\arguments{
\item{quantile_forecasts}{tidy data frame of
quantile forecasts to format,
which should have a location column, a column of
quantile values, a column of quantile levels,
and columns indicating the target end date.}

\item{reference_date}{reference date for the hub forecast.}

\item{target_name}{Name of the target for the table.}

\item{quantile_value_col}{name of the column containing the quantile
values. Default \code{'quantile_value'} (the default output
column name from \code{\link[=trajectories_to_quantiles]{trajectories_to_quantiles()}}).}

\item{quantile_level_col}{name of the column indicating the
quantile level. Default \code{'quantile_level'} (the default output
column name from \code{\link[=trajectories_to_quantiles]{trajectories_to_quantiles()}}).}

\item{location_col}{name of the column containing the location
value. Default \code{'location'}.}

\item{timepoint_col}{names of the columns indicating the target end date.}

\item{target_end_dates}{Vector of target end dates to include,}

\item{horizons}{Vector of forecast horizons to include,
in \code{horizon_timescale}s ahead of the \code{reference_date}.}

\item{horizon_timescale}{Either "days" or "weeks"}

\item{excluded_locations}{locations to exclude from the table
Default c("60", "78") (American Samoa and the US Virgin Islands,
for which FluSight does not currently accept forecasts).}

\item{quantile_tol}{Round quantile level values to this many digits}

\item{reference_dow}{Which day of the week should the reference_date
be, as an integer?}

\item{week_start}{Starting day of the week relative to
Monday, as an integer. See the \code{\link[lubridate:day]{lubridate::wday()}} documentation for details.}
}
\value{
a properly formatted table, as a \code{\link[tibble:tibble]{tibble}}.
}
\description{
Format quantile forecasts for a hubverse submission
}
