% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorize_prism.R
\name{get_prism_cutpoints}
\alias{get_prism_cutpoints}
\title{Get PRISM activity level cutpoints given
disease and location.}
\usage{
get_prism_cutpoints(
  location,
  disease,
  as_of = lubridate::today(),
  signal = lifecycle::deprecated()
)
}
\arguments{
\item{location}{location(s) for which to return the
cutpoints, as a two-letter abbreviation. Use
\link{us_location_recode} with
\code{location_output_format = "abbr"} to convert to this
format.}

\item{disease}{disease(s) for which to return the
cutpoints. One of \code{"ARI"}, \code{"COVID-19"},
\code{"Influenza"}, or \code{"RSV"}, or an array of those
values. NHSN provides no \code{"ARI"} thresholds.}

\item{as_of}{date for which the cutpoints are valid.
Defaults to today.}

\item{signal}{surveillance signal for which to
return the cutpoints. One of \code{"NSSP"} (proportions
of emergency department visits) or \code{"NHSN"} (weekly
hospital admissions per 100k population). If not
given, defaults to \code{"NSSP"} with a deprecation
warning (a future version will require it).}
}
\value{
The cutpoints, as a list of vectors.
}
\description{
Get PRISM activity level cutpoints given
disease and location.
}
