% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorize_prism.R
\name{get_prism_cutpoints}
\alias{get_prism_cutpoints}
\title{Get PRISM activity level cutpoints for given
disease and location}
\usage{
get_prism_cutpoints(location, disease, as_of = lubridate::today())
}
\arguments{
\item{location}{location(s) for which to return the cutpoints, as
a two-letter abbreviation. Use \link{us_location_recode} with
\code{location_output_format = "abbr"} to convert to this format.}

\item{disease}{disease(s) for which to return the cutpoints.
One of \code{"ARI"}, \code{"COVID-19"}, \code{"Influenza"}, or \code{"RSV"}, or
an array of those value.}

\item{as_of}{date for which the cutpoints are valid. Defaults to today.}
}
\value{
The cutpoints, as a list of vectors.
}
\description{
Get PRISM activity level cutpoints for given
disease and location
}
\examples{
get_prism_cutpoints("WA", "Influenza")

get_prism_cutpoints(c("US", "WA"), "COVID-19")

get_prism_cutpoints(c("US", "WA"), c("ARI", "RSV"), as.Date("2025-01-01"))
}
