% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorize_prism.R
\name{get_prism_cutpoints}
\alias{get_prism_cutpoints}
\title{Get PRISM activity level cutpoints for given
diseases and locations}
\usage{
get_prism_cutpoints(locations, diseases)
}
\arguments{
\item{locations}{location(s) for which to return the cutpoints.
A location two-letter abbreviation as in the \code{short_name}
column of \link{us_location_table}, or an array
of those abbreviations.}

\item{diseases}{disease(s) for which to return the cutpoints.
One of \code{"ARI"}, \code{"COVID-19"}, \code{"Influenza"}, or \code{"RSV"}, or
an array of those values.}
}
\value{
The cutpoints, as an ordered list of vectors.
}
\description{
Get PRISM activity level cutpoints for given
diseases and locations
}
\examples{
get_prism_cutpoints("WA", "Influenza")

get_prism_cutpoints(c("US", "WA"), "COVID-19")

get_prism_cutpoints(c("US", "WA"), c("ARI", "RSV"))
}
