% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorize_prism.R
\name{get_prism_cutpoints}
\alias{get_prism_cutpoints}
\title{Get PRISM activity level cutpoints for a given
disease and location}
\usage{
get_prism_cutpoints(disease, location)
}
\arguments{
\item{disease}{disease(s) for which to return the cutpoints.
One of \code{"ARI"}, \code{"COVID-19"}, \code{"Influenza"}, or \code{"RSV"}, or
an array of those values.}

\item{location}{location(s) for which to return the cutpoints.
A location two-letter abbreviation as in the \code{short_name}
column of \link{us_location_table}, or an array
of such abbreviations.}

\item{The}{cutpoints, as a single vector or array of vectors.}
}
\description{
Get PRISM activity level cutpoints for a given
disease and location
}
\examples{
get_prism_cutpoints("Influenza", "WA")

get_prism_cutpoints("COVID-19", c("US", "WA"))

get_prism_cutpoints(c("ARI", "RSV"), c("US", "WA"))
}
