% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{ggplot_dt}
\alias{ggplot_dt}
\title{plotting function for spatial data}
\usage{
ggplot_dt(
  dt,
  data_col = colnames(dt)[3],
  mn = NULL,
  discrete_cs = FALSE,
  rr = NULL,
  low = "blue",
  mid = "white",
  high = "red",
  name = data_col,
  midpoint = NULL,
  ...,
  binwidth = NULL,
  bin_midpoint = midpoint,
  colorscale = NULL,
  tol = 0
)
}
\arguments{
\item{dt}{Data table containing the data for plotting.}

\item{data_col}{The name of the column in dt containing the data for plotting.}

\item{mn}{optional plot title}

\item{discrete_cs}{Logical. Should the color scale be discretized?}

\item{rr, low, mid, high, name, midpoint, ...}{Arguments for the color scale, passed on to scale_fill_gradient2/scale_fill_steps2 (depending on whether discrete_cs == TRUE).
rr replaces limits (specifying the range of the color scale) for consistency with the other plotting functions.
Unlike for scale_fill_gradient2, the default midpoint is the center of the data range (or the center of rr, if provided), not 0.
Note that specifying the midpoint can often be a convenient way to force a color scale with only two colors (for example, by setting it
to the minimum or maximum of your data). The ... argument can in particular be used to customize the legend/colorbar using the function \code{guide_colorbar},
see https://ggplot2.tidyverse.org/reference/guide_colourbar.html. Moreover, when \code{discrete_cs == TRUE} you can pass the arguments \code{n.breaks,breaks} to customize the scale.
If you use n.breaks you might also want to set nice.breaks = FALSE, see ?scale_fill_steps2.}

\item{binwidth, bin_midpoint}{only used when \code{discrete_cs == TRUE}. Normally, the breaks for the discrete colorscale are specified by n.breaks (which is not reliable,
since they're adjusted to be 'nice'), or by specifying the breaks explicitly (which is often tedious). This gives you a third option, namely specifying how far the breaks
should be apart, and specifying the centerpoint for one of the bins (default is midpoint, or the center of rr if midpoint is not provided). For example, if your color scale shows
percentages and you'd like 4 categories, the easiest way to specify is \code{binwidth = 25, bin_midpoint = 12.5}.}

\item{colorscale}{This allows to pass any ScaleContinuous object to specify the color scale (for full flexibility).
If this is provided, \code{discrete_cs,rr,low,mid,high,name,midpoint,...} are all ignored.}
}
\value{
a ggplot object.
}
\description{
Plots spatial data from a data.table. The data table needs to contain columns named 'lon' and 'lat'. The grid needs to be regular.
The function runs a crude check whether the data is tempo-spatial. If it is, the time-slice of the first row is used.
}
\author{
Claudio Heinrich
}
