% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{ggplot_dt}
\alias{ggplot_dt}
\title{plotting function for spatial data}
\usage{
ggplot_dt(
  dt,
  data_col = colnames(dt)[3],
  mn = NULL,
  rr = NULL,
  low = "blue",
  mid = "white",
  high = "red",
  name = data_col,
  midpoint = NULL,
  ...,
  colorscale = NULL
)
}
\arguments{
\item{dt}{Data table containing the data for plotting.}

\item{data_col}{The name of the column in dt containing the data for plotting.}

\item{mn}{optional plot title}

\item{rr, low, mid, high, name, midpoint, ...}{Arguments for the color scale, passed on to scale_fill_gradient2.
rr replaces limits (specifying the range of the color scale) for consistency with the other plotting functions.
Unlike for scale_fill_gradient2 the default midpoint is the center of the data range (or the center of rr, if provided), not 0.
Note that specifying the midpoint can often be a convenient way to force a color scale with only two colors (for example, by setting it
to the minimum or maximum of your data). Also note that ... can in particular be used to discretize the colorscale,
usually by providing \code{n.breaks}. All of these arguments are ignored if \code{colorscale} is provided.}

\item{colorscale}{This allows to pass any ScaleContinuous object to specify the color scale (for full flexibility).
If this is provided, \code{rr,low,mid,high,name,midpoint,...} are ignored.}
}
\value{
a ggplot object.
}
\description{
Plots spatial data from a data.table. The data table needs to contain columns named 'lon' and 'lat'. The grid needs to be regular.
The function runs a crude check whether the data is tempo-spatial. If it is, the time-slice of the first row is used.
}
\author{
Claudio Heinrich
}
