% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hubverse_quantiles_to_median_qi.R
\name{hub_quantiles_to_median_qi}
\alias{hub_quantiles_to_median_qi}
\title{Convert a hubverse-format table of quantiles into a
\code{ggdist}-format table of quantiles.}
\usage{
hub_quantiles_to_median_qi(
  hubverse_quantile_table,
  .width = 0.95,
  require_only_quantiles = FALSE,
  require_all_widths = FALSE,
  quantile_tol = 10
)
}
\arguments{
\item{hubverse_quantile_table}{Hubverse-format quantile table,
as an object something coerceible to a \code{\link[tibble:tibble]{tibble}}.}

\item{.width}{Vector of interval width(s) to for which to create rows
in the output. Default \code{0.95}, to align with the default value in
\code{\link[ggdist:point_interval]{ggdist::median_qi()}}.}

\item{require_only_quantiles}{Boolean. Error if the table contains
\code{output_type} values other than \code{"quantile"}? If \code{FALSE}, will
silently filter the table to rows with \code{output_type == 'quantile'}.
Default \code{FALSE}.}

\item{require_all_widths}{Boolean. Error if the output is missing any
rows for any of the requested \code{.width}s. If \code{TRUE}, the function will
error if unless it can calculate all the requested intervals for all
the groups present. Default \code{FALSE}.}
}
\value{
A \code{\link[tibble:tibble]{tibble}} in the output format of
\code{\link[ggdist:point_interval]{ggdist::median_qi()}}
added.
}
\description{
Expects a hubverse-format quantile table (e.g. as created by
\code{\link[=get_hubverse_table]{get_hubverse_table()}}). Treats all columns in that table other
than \code{output_type_id}, \code{output_type}, and \code{value} as grouping / id columns,
and returns them in the output table.
}
\examples{

hub_quantiles_to_median_qi(hubExamples::forecast_outputs,
  .width = c(0.5, 0.8)
)

}
