% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hubverse_to_scorable.R
\name{hub_to_scorable_quantiles}
\alias{hub_to_scorable_quantiles}
\title{Create a table for scoring hub model
quantile forecasts from a local copy of
a hub.}
\usage{
hub_to_scorable_quantiles(hub_path, as_of = "latest", ...)
}
\arguments{
\item{hub_path}{Local path to a hubverse-style
forecast hub.}

\item{...}{keyword arguments passed to
\code{\link[=quantile_table_to_scorable]{quantile_table_to_scorable()}}.}
}
\value{
Scorable table, as the output of
\code{\link[scoringutils:as_forecast_quantile]{scoringutils::as_forecast_quantile()}}.
}
\description{
Requires a local version of the
forecast hub at \code{hub_path}.
}
