% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{hubverse_std_colnames}
\alias{hubverse_std_colnames}
\title{Internal and user-exposed constants for forecasttools.
Names of the standard columns for hubverse tables.
Alias for \link[hubUtils:std_colnames]{hubUtils::std_colnames}.}
\format{
An object of class \code{character} of length 4.
}
\usage{
hubverse_std_colnames
}
\description{
Internal and user-exposed constants for forecasttools.
Names of the standard columns for hubverse tables.
Alias for \link[hubUtils:std_colnames]{hubUtils::std_colnames}.
}
\keyword{datasets}
