% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hubverse_to_scorable.R
\name{hubverse_table_with_obs}
\alias{hubverse_table_with_obs}
\title{Join hubverse forecasts to observed data}
\usage{
hubverse_table_with_obs(
  hubverse_forecast_table,
  observation_table,
  obs_value_column = "value",
  obs_location_column = "location",
  obs_date_column = "date",
  obs_value_name = "observed",
  join = "full"
)
}
\arguments{
\item{hubverse_forecast_table}{Forecasts, as a hubverse-format
\code{\link[tibble:tibble]{tibble}}, for instance as produced by
\code{\link[=get_hubverse_table]{get_hubverse_table()}}, with columns including \code{target_end_date},
and \code{location}.}

\item{observation_table}{observations, as a \code{\link[tibble:tibble]{tibble}}.}

\item{obs_value_column}{Name of the column containing
observed values in the \code{observed} table, as a string.
Default \code{"value"}}

\item{obs_location_column}{Name of the column containing
location values in the \code{observed} table, as a string.
Default \code{"location"}}

\item{obs_date_column}{Name of the column containing
date values in the \code{observed} table, as a string.
Default \code{"date"}}

\item{obs_value_name}{Name for the column of observed values
in the resulting table (since \code{"value"} clashes with the forecast
value column in a standard hubverse table. Default \code{"observed"}.}

\item{join}{Which SQL-style \code{dplyr} \link[dplyr:mutate-joins]{mutating join}
function to use when joining the tables. Options are \code{"full"}
(for \code{\link[dplyr:mutate-joins]{dplyr::full_join()}}), \code{"left"} (\code{\link[dplyr:mutate-joins]{dplyr::left_join()}}),
\code{"right"} (\code{\link[dplyr:mutate-joins]{dplyr::right_join()}}), and \code{"inner"}
(\code{\link[dplyr:mutate-joins]{dplyr::inner_join()}}).
In the join, the hubverse forecast table is the left table
(\code{x}) and the observation table is the right table (\code{y}).
Default \code{"full"} (i.e. keep all forecasts and observations,
even if some forecasts do not have a corresponding observation or
some observations do not have a corresponding forecast).}
}
\value{
A \code{\link[tibble:tibble]{tibble}} with the observed values
added.
}
\description{
Expects forecast output in hubverse format (e.g. as created by
\code{\link[=get_hubverse_table]{get_hubverse_table()}}) and an observed data table with location,
date, and value columns. The names of these columns the observed
data table can be configured; defaults are \code{"location"}, \code{"date"},
and \code{"value"}, respectively (direct correspondence with
standard Hub "target data" tables).
}
