% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_reference_date.R
\name{modify_reference_date}
\alias{modify_reference_date}
\title{Modify Reference Date}
\usage{
modify_reference_date(
  original_hub_tbl,
  horizon_timescale,
  reference_date_transform = identity,
  reference_date_col = "reference_date",
  target_end_date_col = "target_end_date",
  horizon_col = "horizon",
  horizon_timescale_col = "horizon_timescale"
)
}
\arguments{
\item{original_hub_tbl}{A data frame containing the hub table.}

\item{horizon_timescale}{The timescale for the horizon. Must be a valid
argument to \code{\link[=horizons_from_target_end_dates]{horizons_from_target_end_dates()}}.}

\item{reference_date_transform}{A function to transform the reference date.
Default is \code{identity}.}

\item{reference_date_col}{The name of the reference date column. Default is
"reference_date".}

\item{target_end_date_col}{The name of the target end date column. Default is
"target_end_date".}

\item{horizon_col}{The name of the horizon column. Default is "horizon".}

\item{horizon_timescale_col}{The name of the horizon timescale column.
Default is "horizon_timescale".}
}
\value{
A modified data frame with the reference date transformed and
horizons recomputed
}
\description{
Modify Reference Date
}
