% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_hubverse_pointinterval.R
\name{plot_hubverse_pointintervals}
\alias{plot_hubverse_pointintervals}
\title{Plot a set of hubverse-formatted forecast
as a set of pointintervals.}
\usage{
plot_hubverse_pointintervals(
  hubverse_table,
  horizons = NULL,
  point_estimate_quantile = 0.5,
  lower_limit_quantile = 0.025,
  upper_limit_quantile = 0.975,
  location_input_format = "hub",
  location_output_format = "abbr"
)
}
\arguments{
\item{hubverse_table}{Hubverse table, as a \code{\link[tibble:tibble]{tibble}}}

\item{location_output_format}{how to code locations in the output plot.
Permitted formats are \code{"abbr"} (state/territory
or nation two letter USPS abbreviation), \code{"hub"} (
legacy 2-digit FIPS code for states and territories, \code{US}
#' for the USA as a whole), and \code{"long_name"} (full English
jurisdiction names). Default \code{"abbr"}.}

\item{horizon}{horizons to plot, as a vector of integers. If \code{NULL},
plot all available, each in its own facet. Default \code{NULL}.}

\item{locations}{set of locations to plot. If NULL,
all locations are plotted. Otherwise, a vector
of location values to plot, as USPS-style
abbreviations (e.g. \verb{c("US", "AL", "AK"}),
US hubverse submission location codes (
e.g. \code{c("US", 01, 02)}), or full English
jurisdiction names
(e.g. \verb{c("United States, "Alabama", "Alaska")}.
Default \code{NULL}.}

\item{location_column_format}{format of the hubverse table
\code{location} column.
Permitted formats are \code{"abbr"} (state/territory
or nation two letter USPS abbreviation), \code{"hub"} (
legacy 2-digit FIPS code for states and territories, \code{US}
for the USA as a whole), and \code{"long_name"} (full English
jurisdiction names; not recommended). Default \code{"hub"}.}
}
\value{
A ggplot2 plot of the forecasts as pointintervals
}
\description{
Plot a set of hubverse-formatted forecast
as a set of pointintervals.
}
