% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_predicted_observed.R
\name{plot_pred_obs_by_forecast_date}
\alias{plot_pred_obs_by_forecast_date}
\title{Plot predictive medians and prediction
intervals against observed data by forecast date.}
\usage{
plot_pred_obs_by_forecast_date(
  scorable_table,
  horizons = NULL,
  forecast_date_col = "reference_date",
  target_date_col = "target_end_date",
  predicted_col = "predicted",
  observed_col = "observed",
  quantile_level_col = "quantile_level",
  horizon_col = "horizon",
  prediction_interval_width = 0.95,
  facet_columns = NULL,
  x_label = "Date",
  y_label = "Target",
  y_transform = "log10",
  quantile_tol = 10
)
}
\arguments{
\item{scorable_table}{quantile table with observations,
as the output of \code{\link[scoringutils:as_forecast_quantile]{scoringutils::as_forecast_quantile()}},
or as a table coercible to a \code{scoringutils}-ready quantile
forecast via \code{\link[scoringutils:as_forecast_quantile]{scoringutils::as_forecast_quantile()}}.}

\item{horizons}{Forecast horizons to plot. If \code{NULL}, plot all
available horizons. Default \code{NULL}.}

\item{forecast_date_col}{Name of the column in \code{scorable_table}
giving the forecast date or forecast reference date.
Default \code{"reference_date"} (as in hubverse schema).}

\item{target_date_col}{Name of the column in \code{scorable_table}
giving the forecast target date for a given prediction.
Default \code{"target_end_date"} (as in hubverse schema).}

\item{facet_columns}{Columns in \code{scorable_table} by which to
facet the resulting plot. Will always facet by \code{forecast_date_col}.
If \code{NULL}, facet by all the columns that make up the \code{scoringutils}
"forecast unit", as determined by \code{\link[scoringutils:get_forecast_unit]{scoringutils::get_forecast_unit()}},
except for \code{target_date_col} and \code{horizon}. This means that, by default,
two distinct predictions for a given target date and horizon should
not be visualized on the same facet.}
}
\value{
The plot, as a \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object.
}
\description{
Plot predictive medians and prediction
intervals against observed data by forecast date.
}
