% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_quantiles.R
\name{plot_quantile_timeseries}
\alias{plot_quantile_timeseries}
\title{Plot a timeseries of quantiles.}
\usage{
plot_quantile_timeseries(
  data,
  time_column,
  value_column,
  quantile_level_column,
  linewidth = 2,
  pointsize = 4,
  pointcolor = "darkblue",
  linecolor = "darkblue"
)
}
\arguments{
\item{data}{Timeseries of quantiles as tidy data,
with one row per timepoint per quantile level.}

\item{time_column}{Name of the column in \code{data}
containing timepoint values, as a string.}

\item{value_column}{Name of the column
in data containing the timeseries values at the
given quantile levels, as a string.}

\item{quantile_level_column}{Name of the column in \code{data}
containing indicating which quantile level the row contains,
as a string.}

\item{linewidth}{\code{linewidth} parameter passed to
\code{\link[ggplot2:geom_path]{ggplot2::geom_line()}}. Default \code{2}.}

\item{pointsize}{\code{size} parameter passed to
\code{\link[ggplot2:geom_point]{ggplot2::geom_point()}} Default \code{4}.}

\item{pointcolor}{\code{color} parameter passed to
\code{\link[ggplot2:geom_point]{ggplot2::geom_point()}} Default \code{"darkblue"}.}

\item{linecolor}{\code{color} parameter passed to
\code{\link[ggplot2:geom_path]{ggplot2::geom_line()}}. Default \code{"darkblue"}.}
}
\value{
The plot, as a ggplot object.
}
\description{
Plot a timeseries of quantiles.
}
