% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{prism_thresholds}
\alias{prism_thresholds}
\title{PRISM respiratory virus activity level thresholds}
\format{
An object of class \code{array} of dimension 53 x 4 x 6.
}
\source{
<data-raw/prism_thresholds.R>
}
\usage{
prism_thresholds
}
\description{
A multi-dimensional array with PRISM
respiratory virus activity level thresholds.
Dimensions, in order, are \code{location},
\code{disease}, and \code{breaks}.
}
\details{
Values of \code{disease} are \code{Influenza}, \code{COVID-19},
\code{RSV}, and \code{ARI} (acute respiratory infections).

Values of \code{breaks} are \code{prop_very_low},
\code{prop_low}, \code{prop_moderate}, \code{prop_high},
\code{prop_very_high}, and \code{prop_upper_bound}.

Values of \code{location} are US jurisdictions
and the United States as a whole, using
USPS two-letter codes (the values of \code{abbr})
in \link{us_location_table}.
}
\keyword{datasets}
