% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_quantiles_to_median_qi.R
\name{quantile_table_to_median_qi}
\alias{quantile_table_to_median_qi}
\title{Convert a tidy table of quantiles into a
\code{ggdist}-format table of quantile intervals.}
\usage{
quantile_table_to_median_qi(
  quantile_table,
  value_col,
  quantile_level_col,
  .width = NULL,
  require_all_medians = TRUE,
  require_all_widths = TRUE,
  quantile_tol = 10
)
}
\arguments{
\item{quantile_table}{Tidy table of quantiles,
as an object coerceible to a \code{\link[tibble:tibble]{tibble}}. Must
have columns representing the quantile level and the associated
value}

\item{value_col}{Column in the table specifying values.}

\item{quantile_level_col}{Column in the table specifying quantile
levels}

\item{.width}{Vector of interval width(s) to for which to create rows
in the output. If \code{NULL} (the default), use all and only the widths
that can be computed from quantiles available in
\code{quantile_table}, as determined by \code{\link[=get_available_qi_widths]{get_available_qi_widths()}}.}

\item{require_all_medians}{Boolean. Error if the output is missing
any medians? If \code{TRUE}, the function will succeed only if it can find a
median (\code{0.5} quantile) value for all groups present in the input
and otherwise will error. Default \code{TRUE}.}

\item{require_all_widths}{Boolean. Error if the output is missing
any of the requested \code{.width}s for any group? If \code{TRUE}, the
function will succeed only if it can calculate all the requested
interval \code{.width}s for all the groups present in the input and
otherwise will error. Default \code{TRUE}.}

\item{quantile_tol}{Round values in the \code{quantile_level_col} to
this many decimal places, to avoid problems with floating point number
equality comparisons. Passed as the \code{digits} argument to
\code{\link[base:Round]{base::round()}}. Default 10.}
}
\value{
A \code{\link[tibble:tibble]{tibble}} in the output format of
\code{\link[ggdist:point_interval]{ggdist::median_qi()}}
}
\description{
Expects a tidy quantile table (e.g. a hubverse quantile table or the
output of \code{\link[scoringutils:as_forecast_quantile]{scoringutils::as_forecast_quantile()}}.
Treats any columns in that table other than those specifying the
quantile level and the associated value as grouping variables,
and returns them in the output table.
}
\seealso{
\code{\link[=hub_quantiles_to_median_qi]{hub_quantiles_to_median_qi()}}
}
