% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hubverse_to_scoreable.R
\name{quantile_table_to_scoreable}
\alias{quantile_table_to_scoreable}
\title{create a \code{scoringutils}-ready table.}
\usage{
quantile_table_to_scoreable(
  hubverse_quantile_table,
  observation_table,
  obs_value_column = "value",
  obs_location_column = "location",
  obs_date_column = "date"
)
}
\arguments{
\item{hubverse_quantile_table}{quantile forecasts,
as a hubverse-format \code{\link[tibble:tibble]{tibble}}, e.g.
as produced by \code{\link[=get_hubverse_table]{get_hubverse_table()}}, with columns including
\code{location}, \code{target_end_date}, \code{output_type}, \code{output_type_id},
and \code{value}.}

\item{observation_table}{observations, as a \code{\link[tibble:tibble]{tibble}}.}

\item{obs_value_column}{Name of the column containing
observed values in the \code{observed} table, as a string.
Default \code{"value"}}

\item{obs_location_column}{Name of the column containing
location values in the \code{observed} table, as a string.
Default \code{"location"}}

\item{obs_date_column}{Name of the column containing
date values in the \code{observed} table, as a string.
Default \code{"date"}}
}
\value{
A \code{\link[data.table:data.table]{data.table}} for scoring,
as the output of \code{\link[scoringutils:as_forecast_quantile]{scoringutils::as_forecast_quantile()}}.
}
\description{
Expects quantile forecast output in hubverse format
(e.g. as created by \code{\link[=get_hubverse_table]{get_hubverse_table()}})
and an observed data table with location, date, and value columns.
The column names in the observed data table can be configured;
defaults are \code{"location"}, \code{"date"}, and
\code{"value"}, respectively (i.e. direct correspondence with
standard hub target data tables).
}
