% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantiles_to_category_dist.R
\name{quantiles_to_category_cdf}
\alias{quantiles_to_category_cdf}
\alias{quantiles_to_category_pmf}
\title{Get an approximate ordinal cumulative distribution function
or probability mass function from quantiles for a continuous
variable.}
\usage{
quantiles_to_category_cdf(
  quantile_levels,
  values,
  category_cutpoints,
  labels = names(category_cutpoints),
  ...
)

quantiles_to_category_pmf(
  quantile_levels,
  values,
  category_cutpoints,
  labels = names(head(category_cutpoints, -1)),
  ...
)
}
\arguments{
\item{quantile_levels}{Vector of quantile levels, passed as
the \code{ps} argument to \code{\link[distfromq:make_p_fn]{distfromq::make_p_fn()}}.}

\item{values}{Vector of associated values of the random variable
at those quantiles, passed as the \code{qs} argument to
\code{\link[distfromq:make_p_fn]{distfromq::make_p_fn()}}.}

\item{category_cutpoints}{Vector (optionally named) of category
cutpoints that could be supplied to \code{\link[base:cut]{base::cut()}} or similar to
bin the continuous random variable. \code{forecasttools} functions such
as \code{\link[=get_prism_cutpoints]{get_prism_cutpoints()}} supply cutpoints in this format.

To follow \code{\link[base:cut]{base::cut()}} conventions, the cutpoint vector
should be of length \eqn{n+1}, where \eqn{n} is the number
of categories. Entries \eqn{1, ..., n} should be the left endpoints of
of each categorical bin. Entry \eqn{n+1} should be the \emph{right} endpoint
of the final bin, i.e. the upper bound of the underlying continuous
variable's support, or \code{Inf} if there is no upper bound.
Similarly, if there is no lower bound to the underlying variable's
support, entry \eqn{1} should be \code{-Inf}.}

\item{labels}{Labels for the CDF or PMF categories. For CDFs,
the \code{labels} vector should be equal in length to the cutpoints vector.
For PMFs, the labels vector should have one fewer entry than the
cutpoints vector (since the last cutpoint is the right endpoint
of the final bin). If no \code{labels} vector is specified,
the function uses the names of the cutpoints vector itself,
if any, and otherwise returns an unnamed vector. For PMFs,
that means all cutpoint names \emph{except} the last.}

\item{...}{Additional keyword arguments passed to
\code{\link[distfromq:make_p_fn]{distfromq::make_p_fn()}}.}
}
\value{
The values of the approximate CDF or PMF for
each provided category,  as a vector
(a named vector if the category bounds vector is named)
}
\description{
Given quantiles that characterize a continuous random variable
and a set of cutpoints for binning that variable into ordinal
categories, return the approximate CDF or PMF for the ordinal
variable.
}
\details{
Wraps \code{\link[distfromq:make_p_fn]{distfromq::make_p_fn()}} to approximate the continuous
CDF, then applies category cutpoints in the form used by
\code{\link[base:cut]{base::cut()}}.

The category cutpoints must cover the entire support of
the continuous random variable: the first cutpoint should
be the lower bound of the support; the final cutpoint should
be the upper bound of the support.

The function will error if the input quantile values include
values outside this support.

The approximate continuous CDF will be estimated on the unconstrained
real interval \eqn{(-\infty, \infty)} with an appropriate monotonic
transform for the support, picked via \code{\link[=get_transform_to_real_line]{get_transform_to_real_line()}}.
It will then be evaluated against the transformed values of the
cutpoints.
}
\examples{

cutpoints <- get_prism_cutpoints("US", "COVID-19", signal = "NSSP")[[1]]
quantile_levels <- c(0.25, 0.5, 0.75)
values <- c(0.05, 0.2, 0.3)
quantiles_to_category_cdf(
     quantile_levels,
     values,
     cutpoints
)

quantiles_to_category_pmf(
     quantile_levels,
     values,
     cutpoints
)

}
