% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_tabular_file}
\alias{read_tabular_file}
\title{Read a tabular file to a tibble, inferring
format from the file extension if it is not
specified.}
\usage{
read_tabular_file(path_to_file, file_format = NULL, ...)
}
\arguments{
\item{path_to_file}{Path to the file to read.}

\item{file_format}{Format of the file to read.
One of \code{"tsv"}, \code{"csv"}, \code{"parquet"}. If \code{NULL},
will be inferred from the file extension. Default
\code{NULL}.}

\item{...}{Additional keyword arguments passed to the
file reader function, which will be one of
\code{\link[readr:read_delim]{readr::read_csv()}}, \code{\link[readr:read_delim]{readr::read_tsv()}}, and
\code{\link[arrow:read_parquet]{arrow::read_parquet()}}, depending on the \code{file_format}.}
}
\value{
The result of reading in the file, as a
\code{\link[tibble:tibble]{tibble}}.
}
\description{
Read a tabular file to a tibble, inferring
format from the file extension if it is not
specified.
}
