% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{soql_nullable_is_in}
\alias{soql_nullable_is_in}
\title{Add an "is in" statement to a SOQL query, or return the original
query if \code{match_values} is  \code{NULL}.}
\usage{
soql_nullable_is_in(soql_list, column, match_values)
}
\arguments{
\item{soql_list}{A \code{soql} query object, which
can be piped in. If one hasn't been
created yet, use or pipe in \code{\link[soql:soql]{soql::soql()}}.}

\item{column}{The column to filter on}

\item{match_values}{A vector of values that column
must match, or \code{NULL} for no filtering.}
}
\value{
A new \code{\link[soql:soql]{soql::soql()}} object with the filter added,
or simply the input object if \code{match_value} is \code{NULL}.
}
\description{
An is in statement is a \code{\link[soql:soql_queries]{soql::soql_where()}} statement
that requires the values of a given column to match one of the
entries of a vector of \code{match_values}.
}
