% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{soql_nullable_select}
\alias{soql_nullable_select}
\title{Add a \code{\link[soql:soql_queries]{soql::soql_select()}} statement to a query
to select a set of columns \code{columns}, or return
the original query if \code{columns} is \code{NULL}.}
\usage{
soql_nullable_select(soql_list, columns)
}
\arguments{
\item{soql_list}{A \code{soql} query object, which
can be piped in. If one hasn't been
created yet, use or pipe in \code{\link[soql:soql]{soql::soql()}}.}

\item{columns}{The columns to select, or \code{NULL}.}
}
\value{
A new \code{\link[soql:soql]{soql::soql()}} object with the selection statement
added, or the input object if \code{columns} is \code{NULL}.
}
\description{
Add a \code{\link[soql:soql_queries]{soql::soql_select()}} statement to a query
to select a set of columns \code{columns}, or return
the original query if \code{columns} is \code{NULL}.
}
