% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_hubverse.R
\name{target_end_dates_from_horizons}
\alias{target_end_dates_from_horizons}
\title{Get a correctly-formatted set of hubverse
\code{target_end_date} values for the given reference
date and set of horizons.}
\usage{
target_end_dates_from_horizons(reference_date, horizons = -1:3)
}
\arguments{
\item{reference_date}{hub forecasting reference date}

\item{horizons}{Vector of forecast horizons to include,
in weeks ahead of the \code{reference_date}.
Default -1:3 (FluSight and Covidhub 2024/25 horizons).}
}
\value{
a \code{\link[tibble:tibble]{tibble}} with columns \code{reference_date},
\code{horizon}, \code{target_end_date},
\code{epiweek}, and \code{epiyear}.
}
\description{
Get a correctly-formatted set of hubverse
\code{target_end_date} values for the given reference
date and set of horizons.
}
