% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_forecasttools.R
\name{theme_forecasttools}
\alias{theme_forecasttools}
\title{forecasttools ggplot2 theme.}
\usage{
theme_forecasttools(...)
}
\arguments{
\item{...}{arguments passed to \code{\link[ggplot2:theme]{ggplot2::theme()}}}
}
\value{
The theme, which can be added to a ggplot
object.
}
\description{
Built on top of \code{\link[ggplot2:ggtheme]{ggplot2::theme_minimal()}}
}
