% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hub_utils.R
\name{update_hub}
\alias{update_hub}
\title{Update the forecast hub located at the path \code{hub_path}.}
\usage{
update_hub(hub_path)
}
\arguments{
\item{hub_path}{Path to forecast hub.}
}
\description{
Update the forecast hub located at the path \code{hub_path}.
}
