% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode_locations.R
\name{us_loc_abbr_to_code}
\alias{us_loc_abbr_to_code}
\title{Convert a two-letter USA location abbreviation to a
two-character USA location code}
\usage{
us_loc_abbr_to_code(abbr)
}
\arguments{
\item{abbr}{vector of USPS two letter name abbreviations}
}
\value{
vector of the same length recoded as hub-style
location codes
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
\description{
Given a vector of state/territory two-letter
USPS short names (e.g. MA, TX, PR), return
the corresponding location code typically used
in forecast hubs (legacy FIPS code for states and territories,
\code{US} for the US).
}
\seealso{
\code{\link[=us_location_lookup]{us_location_lookup()}}
}
