% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode_locations.R
\name{us_location_lookup}
\alias{us_location_lookup}
\alias{location_lookup}
\title{Look up rows of the USA location table
corresponding to the entries of a given
vector.}
\usage{
us_location_lookup(
  location,
  location_input_format,
  location_output_format = NULL
)

location_lookup(location, location_input_format, location_output_format = NULL)
}
\arguments{
\item{location}{vector of location values}

\item{location_input_format}{format in which the location
vector is coded.
Permitted formats are \code{"abbr"} (state/territory
or nation two letter USPS abbreviation), \code{"hub"}
(legacy 2-digit FIPS code for states and territories, \code{US}
for the USA as a whole), and \code{"long_name"} (full English
name for jurisdiction).}

\item{location_output_format}{Return only this column of the
output table, if it is provided. Otherwise return the whole
table. Default \code{NULL} (return all columns).
or nation two letter USPS abbreviation), \code{"hub"}
(legacy 2-digit FIPS code for states and territories, \code{US}
for the USA as a whole), and \code{"long_name"} (full English
name for jurisdiction).}
}
\value{
the corresponding rows of the \link{us_location_table}
matching the location vector, with repeats possible, or the
values of those rows for a given column, as specified in
\code{location_output_format}.
}
\description{
Get the rows of
\link{us_location_table}
corresponding to a given location vector
and format, with repeats possible
}
