% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{us_location_pop}
\alias{us_location_pop}
\title{Population estimates for US locations}
\format{
A tibble with 53 rows and 2 columns:
\describe{
\item{name}{Full name of the location}
\item{population}{Estimated population count}
}
}
\source{
Census Population Estimates Program via \code{tidycensus::get_estimates()}

data-raw/us_location_pop.R
}
\usage{
us_location_pop
}
\description{
A table of population estimates for the United States as a whole
and for US states, using the most recent Census Population Estimates
Program vintage available when the dataset was generated.
}
\keyword{datasets}
