% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hubverse_quantiles_to_median_qi.R
\name{widths_to_qi_table}
\alias{widths_to_qi_table}
\title{Get a table of quantile intervals ("qi") corresponding to a
vector of interval widths.}
\usage{
widths_to_qi_table(widths, quantile_tol = 10)
}
\arguments{
\item{widths}{Vector of interval widths. All entries
must be between 0 and 1, inclusive. Duplicate entries
are ignored.}

\item{quantile_tol}{Round quantile level values to this many
decimal places, to avoid problems with floating point number
equality comparisons. Passed as the \code{digits} argument to
\code{\link[base:Round]{base::round()}}. Default 10.}
}
\value{
A table of available intervals, as a
\code{\link[tibble:tibble]{tibble}}, with columns
\code{.lower_quantile}, \code{.upper_quantile}, and \code{.width},
that give the bounds and the widths of the requested
intervals.
}
\description{
These intervals are also called "equal-tailed intervals", and can
be computed whenever both the \code{x} quantile and the \code{1 - x} are present.
(yielding an interval of width \code{1 - (min(x, 1 - x) / 2)}.
}
